Separate the represented quest responsibility from the Session root (#605)
Fifth delivery of the refactor push. 124 methods move into 7 bounded
submodules under session/quest: state, objectives, rewards, giver,
publication, persistence and reset. Every file is under the 1,000-line review
budget.

session/mod.rs drops from 50,802 to 47,905 production lines. Across #597,
#599, #601, #603 and #605 the root falls from 75,841 to 47,905 - 37 per cent
retired into navigable owners.

Quest was chosen by comparing scoped candidates rather than raw keyword
counts: quest 2,706 lines, pet 2,686, persistence 2,100, progression 1,948,
combat 1,477, social 1,461. Methods that merely mention a quest but belong to
creature, gameobject, item or loot responsibilities - several already
separated by #597 and #599 - stay where they are.

Safety established by comparison: the impl surface is 1,577 methods before and
after, an identical multiset, and the ownership baseline reports 3,761
associated items unchanged. Seventeen baseline differences are private to
pub(in crate::session) and nothing became public. Two further differences are
a trailing comma in signatures rustfmt wrapped during #603, with the
visibility unchanged - cosmetic, not semantic. The two direct-registry deltas
cover the identical method set, re-attributed to their new files.

Verified: cargo check -p wow-world --lib --all-targets clean, cargo test
-p wow-world --lib 3,822 passed and 0 failed, session-ownership-check
--syntax-only PASS, check_architecture.py check and self-test PASS, fmt and
diff hygiene clean. Ceiling tightened to the validated figure.
