+++ crates/wow-world/src/session/mod.rs
@@ -20,6 +20,7 @@ pub mod mailbox;
 mod movement;
 mod player_cast;
 mod player_items;
+mod quest;
 pub mod registry;
 mod spell_state;
 mod trainer_acquisition;
@@ -9600,39 +9601,6 @@ impl WorldSession {
         })
     }
 
-    pub fn set_game_event_quest_complete_sender_like_cpp(
-        &mut self,
-        sender: flume::Sender<GameEventQuestCompleteCommandLikeCpp>,
-    ) {
-        self.game_event_quest_complete_tx = Some(sender);
-    }
-
-    pub async fn notify_game_event_quest_complete_like_cpp(
-        &self,
-        quest_id: u32,
-    ) -> GameEventQuestCompleteClientOutcomeLikeCpp {
-        let Some(sender) = self.game_event_quest_complete_tx.as_ref() else {
-            return GameEventQuestCompleteClientOutcomeLikeCpp::SenderMissing { quest_id };
-        };
-
-        let (response_tx, response_rx) = flume::bounded(1);
-        let command = GameEventQuestCompleteCommandLikeCpp {
-            quest_id,
-            response_tx,
-        };
-        if sender.try_send(command).is_err() {
-            return GameEventQuestCompleteClientOutcomeLikeCpp::SendFailed { quest_id };
-        }
-
-        match tokio::time::timeout(Duration::from_millis(250), response_rx.recv_async()).await {
-            Ok(Ok(response)) => GameEventQuestCompleteClientOutcomeLikeCpp::Ok(response),
-            Ok(Err(_)) => {
-                GameEventQuestCompleteClientOutcomeLikeCpp::ResponseChannelClosed { quest_id }
-            }
-            Err(_) => GameEventQuestCompleteClientOutcomeLikeCpp::ResponseTimeout { quest_id },
-        }
-    }
-
     #[cfg(test)]
     fn initial_player_fixture_like_cpp(&self) -> Option<Player> {
         self.build_initial_player_for_owner_like_cpp(
@@ -12930,46 +12898,6 @@ impl WorldSession {
         [source.loot_id, source.personal_loot_id, source.push_loot_id]
     }
 
-    fn represented_current_player_has_incomplete_quest_objective_for_object_id_like_cpp(
-        &self,
-        object_id: i32,
-    ) -> bool {
-        let Some(quest_store) = self.quest_store.as_ref() else {
-            return false;
-        };
-        let Some(quests) = self.player_quest_gameplay_snapshot_like_cpp() else {
-            return false;
-        };
-        quests.statuses.values().any(|status| {
-            if status.status != crate::conditions::QUEST_STATUS_INCOMPLETE_LIKE_CPP {
-                return false;
-            }
-            let Some(quest) = quest_store.get(status.quest_id) else {
-                return false;
-            };
-            quest
-                .objectives
-                .iter()
-                .enumerate()
-                .any(|(fallback_index, objective)| {
-                    if objective.obj_type != QUEST_OBJECTIVE_ITEM_LIKE_CPP
-                        || objective.object_id != object_id
-                    {
-                        return false;
-                    }
-                    let storage_index = usize::try_from(objective.storage_index)
-                        .ok()
-                        .unwrap_or(fallback_index);
-                    let current = status
-                        .objective_counts
-                        .get(storage_index)
-                        .copied()
-                        .unwrap_or(0);
-                    current < objective.amount.max(1)
-                })
-        })
-    }
-
     fn represented_gameobject_is_for_quests_like_cpp(
         &self,
         gameobject_entry: u32,
@@ -14104,19 +14032,6 @@ impl WorldSession {
         self.persistence_ports_like_cpp.world.map_corpse.as_ref()
     }
 
-    pub fn set_quest_poi_persistence_port_like_cpp(
-        &mut self,
-        port: Arc<dyn wow_persistence::QuestPoiPersistencePortLikeCpp>,
-    ) {
-        self.persistence_ports_like_cpp.catalogs.quest_poi = Some(port);
-    }
-
-    pub(crate) fn quest_poi_persistence_port_like_cpp(
-        &self,
-    ) -> Option<Arc<dyn wow_persistence::QuestPoiPersistencePortLikeCpp>> {
-        self.persistence_ports_like_cpp.catalogs.quest_poi.clone()
-    }
-
     pub fn set_group_loot_money_persistence_port_like_cpp(
         &mut self,
         port: Arc<dyn wow_persistence::GroupLootMoneyPersistencePortLikeCpp>,
@@ -14758,110 +14673,6 @@ impl WorldSession {
             currency_id,
             quantity: currency.quantity,
             amount,
-            weekly_quantity: ((currency.weekly_quantity / scaler) > 0)
-                .then_some(currency.weekly_quantity),
-            max_quantity: (max_quantity != 0).then_some(max_quantity),
-            total_earned: entry.has_total_earned().then_some(currency.earned_quantity),
-            suppress_chat_log: entry.is_suppressing_chat_log(false),
-        };
-        if !self.set_player_currencies_like_cpp(currencies) {
-            return Err(());
-        }
-        Ok(Some(delta))
-    }
-
-    /// C++ `Player::AddCurrency(..., CurrencyGainSource::*QuestReward*)`.
-    ///
-    /// This represented seam intentionally fails closed for award conditions and reputation
-    /// conversion currencies until those runtime systems are available to quest rewards.
-    pub(crate) fn add_currency_quest_reward_like_cpp(
-        &mut self,
-        currency_id: u32,
-        amount: u32,
-        gain_source: CurrencyGainSourceLikeCpp,
-    ) -> Result<Option<PlayerCurrencyDelta>, ()> {
-        if amount == 0 {
-            return Ok(None);
-        }
-
-        let Some(entry) = self
-            .currency_types_store
-            .as_ref()
-            .and_then(|store| store.get(currency_id))
-            .copied()
-        else {
-            return Err(());
-        };
-
-        let player_team = player_team_for_race_cpp(self.player_race_like_cpp());
-        if (entry.is_alliance() && player_team != Team::Alliance)
-            || (entry.is_horde() && player_team != Team::Horde)
-        {
-            return Ok(None);
-        }
-
-        if entry.award_condition_id != 0 {
-            return Err(());
-        }
-        if entry.faction_id != 0 || currency_id == CurrencyTypes::Azerite as u32 {
-            return Ok(None);
-        }
-
-        let ignore_caps = matches!(
-            gain_source,
-            CurrencyGainSourceLikeCpp::QuestRewardIgnoreCaps
-                | CurrencyGainSourceLikeCpp::WorldQuestRewardIgnoreCaps
-        );
-        let mut currencies = self.player_currencies_like_cpp().ok_or(())?;
-        let currency = currencies.entry(currency_id).or_insert(PlayerCurrency {
-            state: PlayerCurrencyState::New,
-            quantity: 0,
-            weekly_quantity: 0,
-            tracked_quantity: 0,
-            increased_cap_quantity: 0,
-            earned_quantity: 0,
-            flags: 0,
-        });
-
-        let weekly_cap = entry.max_earnable_per_week;
-        let mut applied = amount;
-        if !ignore_caps {
-            if weekly_cap != 0 && currency.weekly_quantity.saturating_add(applied) > weekly_cap {
-                applied = weekly_cap.saturating_sub(currency.weekly_quantity);
-            }
-
-            let max_quantity = currency_max_quantity_cpp(&entry, currency);
-            if max_quantity != 0 && currency.quantity.saturating_add(applied) > max_quantity {
-                applied = max_quantity.saturating_sub(currency.quantity);
-            }
-        }
-
-        if applied == 0 {
-            return Ok(None);
-        }
-
-        if currency.state != PlayerCurrencyState::New {
-            currency.state = PlayerCurrencyState::Changed;
-        }
-        currency.quantity = currency.quantity.saturating_add(applied);
-        if !ignore_caps {
-            if weekly_cap != 0 {
-                currency.weekly_quantity = currency.weekly_quantity.saturating_add(applied);
-            }
-            if entry.is_tracking_quantity() {
-                currency.tracked_quantity = currency.tracked_quantity.saturating_add(applied);
-            }
-            if entry.has_total_earned() {
-                currency.earned_quantity = currency.earned_quantity.saturating_add(applied);
-            }
-        }
-
-        let scaler = entry.scaler().max(1) as u32;
-        let max_quantity = currency_max_quantity_cpp(&entry, currency);
-        let delta = PlayerCurrencyDelta {
-            currency_id,
-            quantity: currency.quantity,
-            amount: applied,
             weekly_quantity: ((currency.weekly_quantity / scaler) > 0)
                 .then_some(currency.weekly_quantity),
             max_quantity: (max_quantity != 0).then_some(max_quantity),
@@ -18021,28 +17832,6 @@ impl WorldSession {
         }
     }
 
-    pub(crate) fn represented_raid_difficulty_request_like_cpp(
-        &self,
-        difficulty_id: i32,
-        legacy: bool,
-    ) -> Option<u32> {
-        let difficulty_id = u32::try_from(difficulty_id).ok()?;
-        let entry = self
-            .difficulty_store()
-            .and_then(|store| store.get(difficulty_id))
-            .copied()?;
-        if entry.instance_type != MAP_RAID_LIKE_CPP {
-            return None;
-        }
-
-        let flags = DifficultyFlags::from_bits_truncate(entry.flags);
-        if !flags.contains(DifficultyFlags::CAN_SELECT) {
-            return None;
-        }
-
-        (flags.contains(DifficultyFlags::LEGACY) == legacy).then_some(difficulty_id)
-    }
-
     fn set_represented_group_difficulty_like_cpp(
         &mut self,
         difficulty_id: u32,
@@ -18645,33 +18434,6 @@ impl WorldSession {
         self.num_talents_at_level_store.as_ref()
     }
 
-    pub(crate) fn begin_player_quest_status_authority_load_like_cpp(&mut self) {
-        let _ = self.mutate_player_quest_gameplay_like_cpp(|state| {
-            state.status_authority_complete = false;
-            state.rewarded_quest_rows.clear();
-        });
-        self.invalidate_canonical_player_spell_hit_aura_authority_like_cpp();
-    }
-
-    pub(crate) fn record_represented_rewarded_quest_row_like_cpp(&mut self, quest_id: u32) {
-        let _ = self.mutate_player_quest_gameplay_like_cpp(|state| {
-            state.rewarded_quest_rows.insert(quest_id);
-        });
-    }
-
-    pub(crate) fn complete_player_quest_status_authority_load_like_cpp(&mut self) {
-        let _ = self.mutate_player_quest_gameplay_like_cpp(|state| {
-            state.status_authority_complete = true;
-        });
-    }
-
-    pub(crate) fn invalidate_player_quest_status_authority_like_cpp(&mut self) {
-        let _ = self.mutate_player_quest_gameplay_like_cpp(|state| {
-            state.status_authority_complete = false;
-        });
-        self.invalidate_canonical_player_spell_hit_aura_authority_like_cpp();
-    }
-
     pub(crate) fn begin_represented_trait_config_authority_load_like_cpp(&mut self) {
         let _ = self.mutate_player_spell_runtime_like_cpp(
             wow_entities::PlayerSpellRuntimeState::begin_trait_config_load_like_cpp,
@@ -18739,214 +18501,6 @@ impl WorldSession {
             runtime.trait_entry_rows_empty = entries_empty;
         })
         .is_some()
-    }
-
-    pub(crate) fn player_quest_gameplay_snapshot_like_cpp(
-        &self,
-    ) -> Option<PlayerQuestGameplayState> {
-        #[cfg(test)]
-        if self.player_handle_like_cpp.is_none() {
-            return Some(self.player_quest_gameplay_fixture_like_cpp());
-        }
-        self.with_owned_player_like_cpp(|player| player.gameplay_state().quests.clone())
-    }
-
-    pub(crate) fn represented_player_quest_status_like_cpp(
-        &self,
-        quest_id: u32,
-    ) -> Option<Option<u8>> {
-        let state = self.player_quest_gameplay_snapshot_like_cpp()?;
-        Some(state.statuses.get(&quest_id).map(|status| status.status))
-    }
-
-    pub(crate) fn represented_player_has_rewarded_quest_like_cpp(
-        &self,
-        quest_id: u32,
-    ) -> Option<bool> {
-        Some(
-            self.player_quest_gameplay_snapshot_like_cpp()?
-                .rewarded_quest_ids
-                .contains(&quest_id),
-        )
-    }
-
-    pub(crate) fn mutate_player_quest_gameplay_like_cpp<R>(
-        &mut self,
-        mutate: impl FnOnce(&mut PlayerQuestGameplayState) -> R,
-    ) -> Option<R> {
-        let mut mutate = Some(mutate);
-        #[cfg(test)]
-        if self.player_handle_like_cpp.is_none() {
-            let mut fixture = self.player_quest_gameplay_fixture_like_cpp();
-            let result = mutate.take().expect("test quest mutation executes once")(&mut fixture);
-            self.apply_player_quest_gameplay_fixture_like_cpp(fixture);
-            return Some(result);
-        }
-        let canonical = self.mutate_canonical_player_like_cpp(|player| {
-            mutate.take().expect("Player quest mutation executes once")(
-                &mut player.gameplay_state_mut().quests,
-            )
-        });
-        if canonical.is_some() {
-            #[cfg(test)]
-            if let Some(state) =
-                self.with_owned_player_like_cpp(|player| player.gameplay_state().quests.clone())
-            {
-                self.apply_player_quest_core_compatibility_like_cpp(&state);
-            }
-            return canonical;
-        }
-        None
-    }
-
-    #[cfg(test)]
-    fn player_quest_gameplay_fixture_like_cpp(&self) -> PlayerQuestGameplayState {
-        let objective_counts_by_quest = self
-            .player_quests
-            .values()
-            .map(|status| (status.quest_id, status.objective_counts.clone()))
-            .collect();
-        PlayerQuestGameplayState {
-            statuses: self
-                .player_quests
-                .iter()
-                .map(|(&quest_id, status)| (quest_id, status.clone()))
-                .collect(),
-            rewarded_quest_ids: self.rewarded_quests.iter().copied().collect(),
-            daily_quest_ids: self
-                .daily_quests_completed_like_cpp
-                .iter()
-                .copied()
-                .collect(),
-            weekly_quest_ids: self
-                .weekly_quests_completed_like_cpp
-                .iter()
-                .copied()
-                .collect(),
-            monthly_quest_ids: self
-                .monthly_quests_completed_like_cpp
-                .iter()
-                .copied()
-                .collect(),
-            seasonal_quests: self.seasonal_quests_like_cpp.clone(),
-            df_quest_ids: self.df_quests_like_cpp.iter().copied().collect(),
-            last_daily_quest_time_secs: self.last_daily_quest_time_like_cpp,
-            seasonal_quest_changed: self.seasonal_quest_changed_like_cpp,
-            status_authority_complete: self.player_quest_status_authority_complete_like_cpp,
-            rewarded_quest_rows: self.represented_rewarded_quest_rows_like_cpp.clone(),
-            objective_counts_by_quest,
-            ..Default::default()
-        }
-    }
-
-    #[cfg(test)]
-    fn apply_player_quest_gameplay_fixture_like_cpp(&mut self, state: PlayerQuestGameplayState) {
-        self.apply_player_quest_core_compatibility_like_cpp(&state);
-        self.daily_quests_completed_like_cpp = state.daily_quest_ids.into_iter().collect();
-        self.weekly_quests_completed_like_cpp = state.weekly_quest_ids.into_iter().collect();
-        self.monthly_quests_completed_like_cpp = state.monthly_quest_ids.into_iter().collect();
-        self.seasonal_quests_like_cpp = state.seasonal_quests;
-        self.df_quests_like_cpp = state.df_quest_ids.into_iter().collect();
-        self.last_daily_quest_time_like_cpp = state.last_daily_quest_time_secs;
-        self.seasonal_quest_changed_like_cpp = state.seasonal_quest_changed;
-    }
-
-    #[cfg(test)]
-    fn apply_player_quest_core_compatibility_like_cpp(&mut self, state: &PlayerQuestGameplayState) {
-        self.player_quests = state
-            .statuses
-            .iter()
-            .map(|(&quest_id, status)| (quest_id, status.clone()))
-            .collect();
-        self.rewarded_quests = state.rewarded_quest_ids.iter().copied().collect();
-        self.player_quest_status_authority_complete_like_cpp = state.status_authority_complete;
-        self.represented_rewarded_quest_rows_like_cpp = state.rewarded_quest_rows.clone();
-    }
-
-    pub(crate) fn represented_quest_can_increase_rewarded_counters_like_cpp(
-        &self,
-        quest_id: u32,
-    ) -> Option<bool> {
-        self.quest_store.as_ref()?.get(quest_id).map(|quest| {
-            !quest.is_df_quest_like_cpp()
-                && !quest.is_daily_like_cpp()
-                && (!quest.is_repeatable()
-                    || quest.is_weekly_like_cpp()
-                    || quest.is_monthly_like_cpp()
-                    || quest.is_seasonal_like_cpp())
-        })
-    }
-
-    fn represented_spell_area_quest_status_like_cpp(&self, quest_id: u32) -> Option<u8> {
-        let state = self.player_quest_gameplay_snapshot_like_cpp()?;
-        if !state.status_authority_complete {
-            return None;
-        }
-
-        if state.rewarded_quest_ids.contains(&quest_id)
-            && self.represented_quest_can_increase_rewarded_counters_like_cpp(quest_id)?
-        {
-            return Some(crate::conditions::QUEST_STATUS_REWARDED_LIKE_CPP);
-        }
-
-        Some(
-            state
-                .statuses
-                .get(&quest_id)
-                .map(|quest| quest.status)
-                .unwrap_or(crate::conditions::QUEST_STATUS_NONE_LIKE_CPP),
-        )
-    }
-
-    fn represented_quest_login_aura_sources_are_hit_inert_like_cpp(
-        &self,
-        difficulty_id: u8,
-    ) -> bool {
-        let Some(state) = self.player_quest_gameplay_snapshot_like_cpp() else {
-            return false;
-        };
-        if !state.status_authority_complete {
-            return false;
-        }
-        let Some(quests) = self.quest_store.as_ref() else {
-            return state.rewarded_quest_rows.is_empty() && state.statuses.is_empty();
-        };
-
-        // C++ `_LoadQuestStatusRewarded` calls `LearnQuestRewardedSpells`
-        // before filtering special quests out of `m_RewardedQuests`. Keep the
-        // narrow TESTBOT proof to rows whose static reward spell is absent.
-        if state.rewarded_quest_rows.iter().any(|quest_id| {
-            quests
-                .get(*quest_id)
-                .is_none_or(|quest| quest.reward_spell != 0)
-        }) {
-            return false;
-        }
-
-        state.statuses.values().all(|status| {
-            let Some(quest) = quests.get(status.quest_id) else {
-                return false;
-            };
-            let recasts_accept_spell = quest.flags & QUEST_FLAGS_PLAYER_CAST_ACCEPT_LIKE_CPP != 0
-                && quest.flags_ex & QUEST_FLAGS_EX_RECAST_ACCEPT_SPELL_ON_LOGIN_LIKE_CPP != 0
-                && quest.source_spell_id != 0;
-            !recasts_accept_spell
-                || self
-                    .player_target_spell_is_hit_inert_like_cpp(quest.source_spell_id, difficulty_id)
-        })
-    }
-
-    /// C++ `Player::PushQuests` scans the complete global quest-template map
-    /// during login and area updates. `AddQuest` can cast an AUTO_PUSH quest's
-    /// SourceSpellID, so the narrow empty-source proof requires that the
-    /// authoritative store contain no such template at all.
-    fn represented_auto_push_quest_aura_source_is_empty_like_cpp(&self) -> bool {
-        const QUEST_FLAGS_EX_AUTO_PUSH_LIKE_CPP: u32 = 0x0400_0000;
-        self.quest_store.as_ref().is_some_and(|quests| {
-            quests
-                .quests_like_cpp()
-                .all(|quest| quest.flags_ex & QUEST_FLAGS_EX_AUTO_PUSH_LIKE_CPP == 0)
-        })
     }
 
     /// C++ can load/summon a `character_pet` during the Player lifetime and
@@ -19108,26 +18662,6 @@ impl WorldSession {
         self.spell_pet_aura_store
             .as_ref()
             .and_then(|store| store.get_pet_aura_like_cpp(spell_id, effect_index))
-    }
-
-    pub(crate) fn spell_area_for_quest_map_bounds_like_cpp(
-        &self,
-        quest_id: u32,
-    ) -> Vec<&SpellAreaLikeCpp> {
-        self.spell_area_store
-            .as_ref()
-            .map(|store| store.spell_area_for_quest_map_bounds_like_cpp(quest_id))
-            .unwrap_or_default()
-    }
-
-    pub(crate) fn spell_area_for_quest_end_map_bounds_like_cpp(
-        &self,
-        quest_id: u32,
-    ) -> Vec<&SpellAreaLikeCpp> {
-        self.spell_area_store
-            .as_ref()
-            .map(|store| store.spell_area_for_quest_end_map_bounds_like_cpp(quest_id))
-            .unwrap_or_default()
     }
 
     #[cfg(test)]
@@ -19583,42 +19117,6 @@ impl WorldSession {
 
     pub(crate) fn fishing_base_skill_store(&self) -> Option<&Arc<FishingBaseSkillStoreLikeCpp>> {
         self.fishing_base_skill_store.as_ref()
-    }
-
-    /// Set the quest store shared reference.
-    pub fn set_quest_store(&mut self, store: Arc<wow_data::quest::QuestStore>) {
-        self.invalidate_canonical_player_spell_hit_aura_authority_like_cpp();
-        self.quest_store = Some(store);
-    }
-
-    /// Set the represented QuestPoolMgr active snapshot shared reference.
-    pub fn set_quest_pool_store(&mut self, store: Arc<wow_data::quest::QuestPoolStoreLikeCpp>) {
-        self.quest_pool_store = Some(store);
-    }
-
-    /// Set the QuestV2 store shared reference used for C++ quest unique-bit lookups.
-    pub fn set_quest_v2_store(&mut self, store: Arc<QuestV2Store>) {
-        self.quest_v2_store = Some(store);
-    }
-
-    /// Set the QuestInfo store used by C++ Quest::GetQuestTag/IsImportant.
-    pub fn set_quest_info_store(&mut self, store: Arc<QuestInfoStore>) {
-        self.quest_info_store = Some(store);
-    }
-
-    /// Set C++ `CONFIG_QUEST_LOW_LEVEL_HIDE_DIFF`.
-    pub fn set_quest_low_level_hide_diff_like_cpp(&mut self, value: u32) {
-        self.quest_low_level_hide_diff_like_cpp = value;
-    }
-
-    /// Set C++ `CONFIG_QUEST_HIGH_LEVEL_HIDE_DIFF`.
-    pub fn set_quest_high_level_hide_diff_like_cpp(&mut self, value: u32) {
-        self.quest_high_level_hide_diff_like_cpp = value;
-    }
-
-    /// Set the QuestFactionReward store used by C++ quest reputation reward lookup.
-    pub fn set_quest_faction_reward_store(&mut self, store: Arc<QuestFactionRewardStore>) {
-        self.quest_faction_reward_store = Some(store);
     }
 
     async fn reconcile_durable_loot_money_before_save_like_cpp(&mut self) -> bool {
@@ -21206,54 +20704,6 @@ impl WorldSession {
             })
     }
 
-    fn resolved_current_player_xp_persistence_request_like_cpp(
-        &self,
-        level_changed: bool,
-        rest_info_changed: bool,
-        guid_counter: u64,
-    ) -> Option<wow_persistence::PlayerXpPersistenceRequestLikeCpp> {
-        let rest = if rest_info_changed {
-            Some(wow_persistence::PlayerXpRestStateSaveLikeCpp {
-                rest_state: self.resolved_xp_rest_state_like_cpp()?,
-                player_flags: self.resolved_player_flags_for_rest_state_save_like_cpp()?,
-                rest_bonus: Self::sanitize_rest_bonus_like_cpp(
-                    self.resolved_xp_rest_bonus_like_cpp()?,
-                ),
-            })
-        } else {
-            None
-        };
-        Some(wow_persistence::PlayerXpPersistenceRequestLikeCpp {
-            player_guid: guid_counter,
-            level_changed,
-            level: self.player_level_like_cpp(),
-            xp: self.resolved_player_xp_like_cpp()?,
-            rest,
-        })
-    }
-
-    #[cfg(test)]
-    fn current_player_xp_persistence_request_like_cpp(
-        &self,
-        level_changed: bool,
-        rest_info_changed: bool,
-        guid_counter: u64,
-    ) -> wow_persistence::PlayerXpPersistenceRequestLikeCpp {
-        wow_persistence::PlayerXpPersistenceRequestLikeCpp {
-            player_guid: guid_counter,
-            level_changed,
-            level: self.player_level_like_cpp(),
-            xp: self.player_xp_like_cpp(),
-            rest: rest_info_changed.then(|| wow_persistence::PlayerXpRestStateSaveLikeCpp {
-                rest_state: self.represented_xp_rest_state_like_cpp(),
-                player_flags: self.represented_player_flags_for_rest_state_save_like_cpp(),
-                rest_bonus: Self::sanitize_rest_bonus_like_cpp(
-                    self.represented_xp_rest_bonus_like_cpp(),
-                ),
-            }),
-        }
-    }
-
     fn represented_xp_rest_info_changed_since_like_cpp(
         &self,
         old_rest_bonus: f32,
@@ -21863,105 +21313,6 @@ impl WorldSession {
         &self.represented_area_zone_criteria_like_cpp
     }
 
-    /// Represented C++ `Player::LearnQuestRewardedSpells`.
-    ///
-    /// C++ casts each rewarded quest's `RewardSpell` only when that spell exists,
-    /// has a missing `SPELL_EFFECT_LEARN_SPELL` trigger, and the first learned
-    /// spell is tied to `SKILL_LINE_ABILITY_REWARDED_FROM_QUEST` when it is not
-    /// already known. This represented slice applies only the resulting direct
-    /// learned-spell side effect; full `CastSpell` runtime semantics remain in
-    /// the spell-system roadmap.
-    pub(crate) fn apply_represented_quest_rewarded_spells_like_cpp(&mut self) -> usize {
-        let Some(quest_store) = self.quest_store.clone() else {
-            return 0;
-        };
-
-        let Some(quests) = self.player_quest_gameplay_snapshot_like_cpp() else {
-            return 0;
-        };
-        let mut quest_ids = quests
-            .rewarded_quest_ids
-            .iter()
-            .copied()
-            .collect::<Vec<_>>();
-        quest_ids.sort_unstable();
-
-        let mut learned = 0usize;
-        for quest_id in quest_ids {
-            let Some(quest) = quest_store.get(quest_id) else {
-                continue;
-            };
-            if quest.reward_spell == u32::MAX && quest.source_spell_id != 0 {
-                let Ok(source_spell_id) = i32::try_from(quest.source_spell_id) else {
-                    continue;
-                };
-                learned += self.remove_represented_auras_due_to_spell_like_cpp(source_spell_id);
-                continue;
-            }
-            for spell_id in self.represented_quest_rewarded_spell_triggers_like_cpp(quest) {
-                self.learn_known_spell_like_cpp(spell_id);
-                learned += 1;
-            }
-        }
-
-        learned
-    }
-
-    fn represented_quest_rewarded_spell_triggers_like_cpp(
-        &self,
-        quest: &wow_data::quest::QuestTemplate,
-    ) -> Vec<i32> {
-        if quest.reward_spell == 0 {
-            return Vec::new();
-        }
-
-        let Ok(reward_spell_id) = i32::try_from(quest.reward_spell) else {
-            return Vec::new();
-        };
-
-        let Some(spell_info) = self
-            .spell_store()
-            .and_then(|store| store.get(reward_spell_id))
-        else {
-            return Vec::new();
-        };
-
-        let missing_learn_triggers = spell_info
-            .effects()
-            .iter()
-            .filter(|effect| {
-                effect.effect == wow_data::spell::spell_effect_types::SPELL_EFFECT_LEARN_SPELL
-                    && effect.effect_trigger_spell > 0
-                    && !self
-                        .known_spells_like_cpp()
-                        .contains(&effect.effect_trigger_spell)
-            })
-            .map(|effect| effect.effect_trigger_spell)
-            .collect::<Vec<_>>();
-
-        if missing_learn_triggers.is_empty() || spell_info.effects().is_empty() {
-            return Vec::new();
-        }
-
-        let learned_0 = spell_info.effects()[0].effect_trigger_spell;
-        if learned_0 > 0
-            && !self.known_spells_like_cpp().contains(&learned_0)
-            && !self.skill_store().is_some_and(|store| {
-                store
-                    .get_skill_line_ability_map_bounds_like_cpp(learned_0)
-                    .iter()
-                    .any(|ability| {
-                        ability.acquire_method
-                            == wow_data::skill::SKILL_LINE_ABILITY_REWARDED_FROM_QUEST_LIKE_CPP
-                    })
-            })
-        {
-            return Vec::new();
-        }
-
-        missing_learn_triggers
-    }
-
     pub(crate) fn apply_represented_first_login_reputation_with_catalogs_like_cpp(
         &mut self,
         player_bootstrap: &PlayerBootstrapCatalogsLikeCpp,
@@ -22238,23 +21589,6 @@ impl WorldSession {
     }
 
     #[cfg(test)]
-    fn represented_quest_rewarded_talent_points_like_cpp(&self) -> Option<u32> {
-        let canonical = self.with_owned_player_like_cpp(|player| {
-            player.gameplay_state().quest_rewarded_talent_points
-        });
-        #[cfg(test)]
-        if canonical.is_none() && self.player_handle_like_cpp.is_none() {
-            return Some(
-                self.represented_quest_reward_talent_points_like_cpp
-                    .iter()
-                    .map(|reward| reward.points)
-                    .sum(),
-            );
-        }
-        canonical
-    }
-
-    #[cfg(test)]
     fn represented_calculate_talents_points_like_cpp(&self) -> Option<u32> {
         let base_points = self
             .num_talents_at_level_store()
@@ -22303,35 +21637,6 @@ impl WorldSession {
         if let Some(points) = _points {
             self.player_character_points_like_cpp = points;
         }
-    }
-
-    pub(crate) fn add_represented_quest_reward_talent_points_like_cpp(
-        &mut self,
-        quest_id: u32,
-        points: u32,
-    ) -> bool {
-        let canonical = self
-            .with_owned_player_mut_like_cpp(|player| {
-                let rewarded = &mut player.gameplay_state_mut().quest_rewarded_talent_points;
-                *rewarded = rewarded.saturating_add(points);
-            })
-            .is_some();
-        if canonical {
-            return true;
-        }
-        #[cfg(test)]
-        if self.player_handle_like_cpp.is_none() {
-            self.represented_quest_reward_talent_points_like_cpp.push(
-                RepresentedQuestRewardTalentPointsLikeCpp {
-                    quest_id,
-                    points,
-                    init_talent_for_level_unrepresented: true,
-                },
-            );
-            return true;
-        }
-        let _ = quest_id;
-        false
     }
 
     pub(crate) fn load_represented_talent_row_like_cpp(
@@ -23183,699 +22488,6 @@ impl WorldSession {
         })
     }
 
-    async fn update_represented_storing_value_quest_objective_progress_like_cpp(
-        &mut self,
-        item_guid_generator: &wow_core::ObjectGuidGenerator,
-        objective_type: u8,
-        object_id: i32,
-        add_count: i32,
-        credit_guid: wow_core::ObjectGuid,
-    ) {
-        self.invalidate_player_quest_status_authority_like_cpp();
-        use wow_packet::packets::quest::{
-            QuestUpdateAddCredit, QuestUpdateAddPvpCredit, QuestUpdateComplete,
-        };
-
-        let Some(store) = self.quest_store.clone() else {
-            return;
-        };
-
-        let victim_team =
-            if objective_type == QUEST_OBJECTIVE_PLAYERKILLS_LIKE_CPP && !credit_guid.is_empty() {
-                self.player_registry
-                    .as_ref()
-                    .and_then(|registry| registry.quest_credit_race(credit_guid))
-                    .map(player_team_for_race_cpp)
-            } else {
-                None
-            };
-        let player_team = player_team_for_race_cpp(self.player_race_like_cpp());
-        let Some(quests) = self.player_quest_gameplay_snapshot_like_cpp() else {
-            return;
-        };
-        let matching: Vec<(u32, usize, i32, u32)> = quests
-            .statuses
-            .values()
-            .filter(|qs| qs.status == crate::conditions::QUEST_STATUS_INCOMPLETE_LIKE_CPP)
-            .flat_map(|qs| {
-                let quest = store.get(qs.quest_id)?;
-                Some(quest.objectives.iter().filter_map(move |obj| {
-                    if obj.obj_type != objective_type || obj.object_id != object_id {
-                        return None;
-                    }
-                    if objective_type == QUEST_OBJECTIVE_PLAYERKILLS_LIKE_CPP
-                        && (obj.flags & QUEST_OBJECTIVE_FLAG_KILL_PLAYERS_SAME_FACTION_LIKE_CPP)
-                            != 0
-                        && victim_team.is_some_and(|team| team != player_team)
-                    {
-                        return None;
-                    }
-                    let Ok(idx) = usize::try_from(obj.storage_index) else {
-                        return None;
-                    };
-                    Some((qs.quest_id, idx, obj.amount, obj.id))
-                }))
-            })
-            .flatten()
-            .collect();
-
-        let mut quests_to_complete = Vec::new();
-        let mut quests_to_save = Vec::new();
-        for (quest_id, obj_idx, required, objective_id) in matching {
-            let Some(current) = self
-                .mutate_player_quest_gameplay_like_cpp(|quests| {
-                    let qs = quests.statuses.get_mut(&quest_id)?;
-                    if qs.objective_counts.len() <= obj_idx {
-                        qs.objective_counts.resize(obj_idx + 1, 0);
-                    }
-                    if add_count >= 0 && qs.objective_counts[obj_idx] >= required {
-                        return None;
-                    }
-                    qs.objective_counts[obj_idx] = qs.objective_counts[obj_idx]
-                        .saturating_add(add_count)
-                        .clamp(0, required);
-                    Some(qs.objective_counts[obj_idx])
-                })
-                .flatten()
-            else {
-                continue;
-            };
-            quests_to_save.push(quest_id);
-
-            debug!(
-                account = self.account_id,
-                quest_id,
-                obj_idx,
-                current,
-                required,
-                objective_type,
-                object_id,
-                "Quest objective progress"
-            );
-
-            if add_count > 0 {
-                if objective_type == QUEST_OBJECTIVE_PLAYERKILLS_LIKE_CPP {
-                    self.send_packet(&QuestUpdateAddPvpCredit {
-                        quest_id,
-                        count: current as u16,
-                    });
-                } else {
-                    self.send_packet(&QuestUpdateAddCredit {
-                        victim_guid: credit_guid,
-                        quest_id,
-                        object_id,
-                        count: current as u16,
-                        required: required as u16,
-                        objective_type,
-                    });
-                }
-            }
-
-            if current >= required {
-                if let (Some(quest), Some(quests)) = (
-                    store.get(quest_id),
-                    self.player_quest_gameplay_snapshot_like_cpp(),
-                ) {
-                    let quest_already_rewarded = quests.rewarded_quest_ids.contains(&quest_id);
-                    if quests.statuses.get(&quest_id).is_some_and(|status| {
-                        Self::represented_can_complete_quest_after_objective_like_cpp(
-                            status,
-                            quest,
-                            objective_id,
-                            quest_already_rewarded,
-                        )
-                    }) {
-                        quests_to_complete.push(quest_id);
-                    }
-                }
-            }
-        }
-
-        for quest_id in quests_to_complete {
-            let Some(quest) = store.get(quest_id).cloned() else {
-                continue;
-            };
-            let completed = self
-                .complete_represented_quest_after_add_with_generator_like_cpp(
-                    item_guid_generator,
-                    &quest,
-                )
-                .await;
-            if completed {
-                if self.represented_player_quest_status_like_cpp(quest_id)
-                    == Some(Some(crate::conditions::QUEST_STATUS_COMPLETE_LIKE_CPP))
-                {
-                    self.send_packet(&QuestUpdateComplete { quest_id });
-                }
-                info!(
-                    account = self.account_id,
-                    quest_id, "Quest objectives complete"
-                );
-            }
-        }
-        self.save_changed_represented_quest_statuses_like_cpp(&mut quests_to_save)
-            .await;
-        self.sync_player_registry_state_like_cpp();
-    }
-
-    async fn update_represented_storing_flag_quest_objective_progress_like_cpp(
-        &mut self,
-        item_guid_generator: &wow_core::ObjectGuidGenerator,
-        objective_type: u8,
-        object_id: i32,
-        add_count: i32,
-    ) {
-        self.invalidate_player_quest_status_authority_like_cpp();
-        use wow_packet::packets::quest::{QuestUpdateAddCreditSimple, QuestUpdateComplete};
-
-        let Some(store) = self.quest_store.clone() else {
-            return;
-        };
-
-        let Some(quests) = self.player_quest_gameplay_snapshot_like_cpp() else {
-            return;
-        };
-        let matching: Vec<(u32, usize, u32)> = quests
-            .statuses
-            .values()
-            .filter(|qs| qs.status == crate::conditions::QUEST_STATUS_INCOMPLETE_LIKE_CPP)
-            .flat_map(|qs| {
-                let quest = store.get(qs.quest_id)?;
-                Some(quest.objectives.iter().filter_map(move |obj| {
-                    if obj.obj_type != objective_type
-                        || obj.object_id != object_id
-                        || !obj.is_storing_flag_like_cpp()
-                    {
-                        return None;
-                    }
-                    let Ok(idx) = usize::try_from(obj.storage_index) else {
-                        return None;
-                    };
-                    Some((qs.quest_id, idx, obj.id))
-                }))
-            })
-            .flatten()
-            .collect();
-
-        let mut quests_to_complete = Vec::new();
-        let mut quests_to_save = Vec::new();
-        for (quest_id, obj_idx, objective_id) in matching {
-            let Some((objective_was_complete, objective_is_now_complete)) = self
-                .mutate_player_quest_gameplay_like_cpp(|quests| {
-                    let qs = quests.statuses.get_mut(&quest_id)?;
-                    if qs.objective_counts.len() <= obj_idx {
-                        qs.objective_counts.resize(obj_idx + 1, 0);
-                    }
-                    let before = qs.objective_counts[obj_idx] != 0;
-                    qs.objective_counts[obj_idx] = i32::from(add_count > 0);
-                    Some((before, qs.objective_counts[obj_idx] != 0))
-                })
-                .flatten()
-            else {
-                continue;
-            };
-            if objective_was_complete != objective_is_now_complete {
-                quests_to_save.push(quest_id);
-            }
-
-            if add_count > 0 {
-                self.send_packet(&QuestUpdateAddCreditSimple {
-                    quest_id,
-                    object_id,
-                    objective_type,
-                });
-            }
-
-            if !objective_was_complete && objective_is_now_complete {
-                if let (Some(quest), Some(quests)) = (
-                    store.get(quest_id),
-                    self.player_quest_gameplay_snapshot_like_cpp(),
-                ) {
-                    let quest_already_rewarded = quests.rewarded_quest_ids.contains(&quest_id);
-                    if quests.statuses.get(&quest_id).is_some_and(|status| {
-                        Self::represented_can_complete_quest_after_objective_like_cpp(
-                            status,
-                            quest,
-                            objective_id,
-                            quest_already_rewarded,
-                        )
-                    }) {
-                        quests_to_complete.push((quest_id, objective_id));
-                    }
-                }
-            }
-        }
-
-        for (quest_id, objective_id) in quests_to_complete {
-            let Some(quest) = store.get(quest_id).cloned() else {
-                continue;
-            };
-            let completed = self
-                .complete_represented_quest_after_objective_with_generator_like_cpp(
-                    item_guid_generator,
-                    &quest,
-                    objective_id,
-                )
-                .await;
-            if completed {
-                if self.represented_player_quest_status_like_cpp(quest_id)
-                    == Some(Some(crate::conditions::QUEST_STATUS_COMPLETE_LIKE_CPP))
-                {
-                    self.send_packet(&QuestUpdateComplete { quest_id });
-                }
-                info!(
-                    account = self.account_id,
-                    quest_id, "Quest objectives complete"
-                );
-            }
-        }
-        self.save_changed_represented_quest_statuses_like_cpp(&mut quests_to_save)
-            .await;
-        self.sync_player_registry_state_like_cpp();
-    }
-
-    async fn update_represented_money_quest_objective_progress_like_cpp(
-        &mut self,
-        item_guid_generator: &wow_core::ObjectGuidGenerator,
-        old_money: u64,
-        new_money: u64,
-    ) {
-        self.invalidate_player_quest_status_authority_like_cpp();
-        use wow_packet::packets::quest::QuestUpdateComplete;
-
-        let Some(store) = self.quest_store.clone() else {
-            return;
-        };
-
-        let old_money = old_money.min(i64::MAX as u64) as i64;
-        let new_money_i64 = new_money.min(i64::MAX as u64) as i64;
-        let add_count = new_money_i64.saturating_sub(old_money);
-        let Some(quests) = self.player_quest_gameplay_snapshot_like_cpp() else {
-            return;
-        };
-        let matching: Vec<(u32, u32, i32, bool, bool)> = quests
-            .statuses
-            .values()
-            .filter(|qs| {
-                qs.status == crate::conditions::QUEST_STATUS_INCOMPLETE_LIKE_CPP
-                    || qs.status == crate::conditions::QUEST_STATUS_COMPLETE_LIKE_CPP
-            })
-            .flat_map(|qs| {
-                let quest = store.get(qs.quest_id)?;
-                Some(quest.objectives.iter().filter_map(move |obj| {
-                    if obj.obj_type != QUEST_OBJECTIVE_MONEY_LIKE_CPP || obj.object_id != 0 {
-                        return None;
-                    }
-                    let objective_was_complete = old_money >= i64::from(obj.amount);
-                    if objective_was_complete && add_count >= 0 {
-                        return None;
-                    }
-                    let objective_is_now_complete =
-                        old_money.saturating_add(add_count) >= i64::from(obj.amount);
-                    Some((
-                        qs.quest_id,
-                        obj.id,
-                        obj.amount,
-                        objective_was_complete,
-                        objective_is_now_complete,
-                    ))
-                }))
-            })
-            .flatten()
-            .collect();
-
-        let mut quests_to_complete = Vec::new();
-        let mut quests_to_save = Vec::new();
-        for (quest_id, objective_id, required, objective_was_complete, objective_is_now_complete) in
-            matching
-        {
-            debug!(
-                account = self.account_id,
-                quest_id,
-                old_money,
-                new_money = new_money_i64,
-                required,
-                objective_was_complete,
-                objective_is_now_complete,
-                "Quest money objective progress"
-            );
-
-            if objective_is_now_complete {
-                if let (Some(quest), Some(quests)) = (
-                    store.get(quest_id),
-                    self.player_quest_gameplay_snapshot_like_cpp(),
-                ) {
-                    let quest_already_rewarded = quests.rewarded_quest_ids.contains(&quest_id);
-                    if quests.statuses.get(&quest_id).is_some_and(|status| {
-                        Self::represented_can_complete_quest_after_objective_like_cpp(
-                            status,
-                            quest,
-                            objective_id,
-                            quest_already_rewarded,
-                        )
-                    }) {
-                        quests_to_complete.push((quest_id, objective_id));
-                    }
-                }
-            } else if objective_was_complete {
-                if self
-                    .mutate_player_quest_gameplay_like_cpp(|quests| {
-                        let Some(status) = quests.statuses.get_mut(&quest_id) else {
-                            return false;
-                        };
-                        if status.status != crate::conditions::QUEST_STATUS_COMPLETE_LIKE_CPP {
-                            return false;
-                        }
-                        status.status = crate::conditions::QUEST_STATUS_INCOMPLETE_LIKE_CPP;
-                        true
-                    })
-                    .unwrap_or(false)
-                {
-                    quests_to_save.push(quest_id);
-                }
-            }
-        }
-
-        for (quest_id, objective_id) in quests_to_complete {
-            let Some(quest) = store.get(quest_id).cloned() else {
-                continue;
-            };
-            let completed = self
-                .complete_represented_quest_after_objective_with_generator_like_cpp(
-                    item_guid_generator,
-                    &quest,
-                    objective_id,
-                )
-                .await;
-            if completed {
-                quests_to_save.push(quest_id);
-                if self.represented_player_quest_status_like_cpp(quest_id)
-                    == Some(Some(crate::conditions::QUEST_STATUS_COMPLETE_LIKE_CPP))
-                {
-                    self.send_packet(&QuestUpdateComplete { quest_id });
-                }
-                info!(
-                    account = self.account_id,
-                    quest_id, "Quest objectives complete"
-                );
-            }
-        }
-        self.save_changed_represented_quest_statuses_like_cpp(&mut quests_to_save)
-            .await;
-        self.sync_player_registry_state_like_cpp();
-    }
-
-    async fn update_represented_currency_quest_objective_progress_like_cpp(
-        &mut self,
-        item_guid_generator: &wow_core::ObjectGuidGenerator,
-        currency_id: u32,
-        change: i32,
-    ) {
-        self.invalidate_player_quest_status_authority_like_cpp();
-        use wow_packet::packets::quest::QuestUpdateComplete;
-
-        let Some(store) = self.quest_store.clone() else {
-            return;
-        };
-
-        let Some(current_quantity) = self.player_currency_quantity(currency_id).map(i64::from)
-        else {
-            return;
-        };
-        let add_count = i64::from(change);
-        let object_id = i32::try_from(currency_id).unwrap_or(i32::MAX);
-        let Some(quests) = self.player_quest_gameplay_snapshot_like_cpp() else {
-            return;
-        };
-        let matching: Vec<(u32, u32, i32, bool, bool)> = quests
-            .statuses
-            .values()
-            .filter(|qs| {
-                qs.status == crate::conditions::QUEST_STATUS_INCOMPLETE_LIKE_CPP
-                    || qs.status == crate::conditions::QUEST_STATUS_COMPLETE_LIKE_CPP
-            })
-            .flat_map(|qs| {
-                let quest = store.get(qs.quest_id)?;
-                Some(quest.objectives.iter().filter_map(move |obj| {
-                    if obj.obj_type != QUEST_OBJECTIVE_CURRENCY_LIKE_CPP
-                        || obj.object_id != object_id
-                    {
-                        return None;
-                    }
-                    let objective_was_complete = current_quantity >= i64::from(obj.amount);
-                    if objective_was_complete && change >= 0 {
-                        return None;
-                    }
-                    let objective_is_now_complete =
-                        current_quantity.saturating_add(add_count) >= i64::from(obj.amount);
-                    Some((
-                        qs.quest_id,
-                        obj.id,
-                        obj.amount,
-                        objective_was_complete,
-                        objective_is_now_complete,
-                    ))
-                }))
-            })
-            .flatten()
-            .collect();
-
-        let mut quests_to_complete = Vec::new();
-        let mut quests_to_save = Vec::new();
-        for (quest_id, objective_id, required, objective_was_complete, objective_is_now_complete) in
-            matching
-        {
-            debug!(
-                account = self.account_id,
-                quest_id,
-                currency_id,
-                current_quantity,
-                change,
-                required,
-                objective_was_complete,
-                objective_is_now_complete,
-                "Quest currency objective progress"
-            );
-
-            if objective_is_now_complete {
-                if let (Some(quest), Some(quests)) = (
-                    store.get(quest_id),
-                    self.player_quest_gameplay_snapshot_like_cpp(),
-                ) {
-                    let quest_already_rewarded = quests.rewarded_quest_ids.contains(&quest_id);
-                    if quests.statuses.get(&quest_id).is_some_and(|status| {
-                        Self::represented_can_complete_quest_after_objective_like_cpp(
-                            status,
-                            quest,
-                            objective_id,
-                            quest_already_rewarded,
-                        )
-                    }) {
-                        quests_to_complete.push((quest_id, objective_id));
-                    }
-                }
-            } else if objective_was_complete {
-                if self
-                    .mutate_player_quest_gameplay_like_cpp(|quests| {
-                        let Some(status) = quests.statuses.get_mut(&quest_id) else {
-                            return false;
-                        };
-                        if status.status != crate::conditions::QUEST_STATUS_COMPLETE_LIKE_CPP {
-                            return false;
-                        }
-                        status.status = crate::conditions::QUEST_STATUS_INCOMPLETE_LIKE_CPP;
-                        true
-                    })
-                    .unwrap_or(false)
-                {
-                    quests_to_save.push(quest_id);
-                }
-            }
-        }
-
-        for (quest_id, objective_id) in quests_to_complete {
-            let Some(quest) = store.get(quest_id).cloned() else {
-                continue;
-            };
-            let completed = self
-                .complete_represented_quest_after_objective_with_generator_like_cpp(
-                    item_guid_generator,
-                    &quest,
-                    objective_id,
-                )
-                .await;
-            if completed {
-                quests_to_save.push(quest_id);
-                if self.represented_player_quest_status_like_cpp(quest_id)
-                    == Some(Some(crate::conditions::QUEST_STATUS_COMPLETE_LIKE_CPP))
-                {
-                    self.send_packet(&QuestUpdateComplete { quest_id });
-                }
-                info!(
-                    account = self.account_id,
-                    quest_id, "Quest objectives complete"
-                );
-            }
-        }
-        self.save_changed_represented_quest_statuses_like_cpp(&mut quests_to_save)
-            .await;
-        self.sync_player_registry_state_like_cpp();
-    }
-
-    async fn update_represented_reputation_quest_objective_progress_like_cpp(
-        &mut self,
-        item_guid_generator: &wow_core::ObjectGuidGenerator,
-        objective_type: u8,
-        faction_id: u32,
-        change: i32,
-    ) {
-        self.invalidate_player_quest_status_authority_like_cpp();
-        use wow_packet::packets::quest::QuestUpdateComplete;
-
-        let Some(store) = self.quest_store.clone() else {
-            return;
-        };
-        let Some(faction_store) = self.faction_store.as_ref() else {
-            return;
-        };
-        let Some(faction_entry) = faction_store.get(faction_id) else {
-            return;
-        };
-
-        let Some(old_reputation) = self.with_reputation_mgr_like_cpp(|mgr| {
-            mgr.reputation_for_faction_like_cpp(
-                faction_entry,
-                self.player_race_like_cpp(),
-                self.player_class_like_cpp(),
-            )
-        }) else {
-            return;
-        };
-        let new_reputation = old_reputation.saturating_add(change);
-        let object_id = i32::try_from(faction_id).unwrap_or(i32::MAX);
-        let Some(quests) = self.player_quest_gameplay_snapshot_like_cpp() else {
-            return;
-        };
-        let matching: Vec<(u32, u32, i32, bool, bool)> = quests
-            .statuses
-            .values()
-            .filter(|qs| {
-                qs.status == crate::conditions::QUEST_STATUS_INCOMPLETE_LIKE_CPP
-                    || qs.status == crate::conditions::QUEST_STATUS_COMPLETE_LIKE_CPP
-            })
-            .flat_map(|qs| {
-                let quest = store.get(qs.quest_id)?;
-                Some(quest.objectives.iter().filter_map(move |obj| {
-                    if obj.obj_type != objective_type || obj.object_id != object_id {
-                        return None;
-                    }
-                    let (objective_was_complete, objective_is_now_complete) = match objective_type {
-                        QUEST_OBJECTIVE_MIN_REPUTATION_LIKE_CPP => {
-                            (old_reputation >= obj.amount, new_reputation >= obj.amount)
-                        }
-                        QUEST_OBJECTIVE_MAX_REPUTATION_LIKE_CPP => {
-                            (old_reputation <= obj.amount, new_reputation <= obj.amount)
-                        }
-                        _ => return None,
-                    };
-                    if objective_was_complete && change >= 0 {
-                        return None;
-                    }
-                    Some((
-                        qs.quest_id,
-                        obj.id,
-                        obj.amount,
-                        objective_was_complete,
-                        objective_is_now_complete,
-                    ))
-                }))
-            })
-            .flatten()
-            .collect();
-
-        let mut quests_to_complete = Vec::new();
-        let mut quests_to_save = Vec::new();
-        for (quest_id, objective_id, required, objective_was_complete, objective_is_now_complete) in
-            matching
-        {
-            debug!(
-                account = self.account_id,
-                quest_id,
-                objective_type,
-                faction_id,
-                old_reputation,
-                new_reputation,
-                required,
-                objective_was_complete,
-                objective_is_now_complete,
-                "Quest reputation objective progress"
-            );
-
-            if objective_is_now_complete {
-                if let (Some(quest), Some(quests)) = (
-                    store.get(quest_id),
-                    self.player_quest_gameplay_snapshot_like_cpp(),
-                ) {
-                    let quest_already_rewarded = quests.rewarded_quest_ids.contains(&quest_id);
-                    if quests.statuses.get(&quest_id).is_some_and(|status| {
-                        Self::represented_can_complete_quest_after_objective_like_cpp(
-                            status,
-                            quest,
-                            objective_id,
-                            quest_already_rewarded,
-                        )
-                    }) {
-                        quests_to_complete.push((quest_id, objective_id));
-                    }
-                }
-            } else if objective_was_complete {
-                if self
-                    .mutate_player_quest_gameplay_like_cpp(|quests| {
-                        let Some(status) = quests.statuses.get_mut(&quest_id) else {
-                            return false;
-                        };
-                        if status.status != crate::conditions::QUEST_STATUS_COMPLETE_LIKE_CPP {
-                            return false;
-                        }
-                        status.status = crate::conditions::QUEST_STATUS_INCOMPLETE_LIKE_CPP;
-                        true
-                    })
-                    .unwrap_or(false)
-                {
-                    quests_to_save.push(quest_id);
-                }
-            }
-        }
-
-        for (quest_id, objective_id) in quests_to_complete {
-            let Some(quest) = store.get(quest_id).cloned() else {
-                continue;
-            };
-            let completed = self
-                .complete_represented_quest_after_objective_with_generator_like_cpp(
-                    item_guid_generator,
-                    &quest,
-                    objective_id,
-                )
-                .await;
-            if completed {
-                quests_to_save.push(quest_id);
-                if self.represented_player_quest_status_like_cpp(quest_id)
-                    == Some(Some(crate::conditions::QUEST_STATUS_COMPLETE_LIKE_CPP))
-                {
-                    self.send_packet(&QuestUpdateComplete { quest_id });
-                }
-                info!(
-                    account = self.account_id,
-                    quest_id, "Quest objectives complete"
-                );
-            }
-        }
-        self.save_changed_represented_quest_statuses_like_cpp(&mut quests_to_save)
-            .await;
-        self.sync_player_registry_state_like_cpp();
-    }
-
     pub(crate) async fn killed_player_credit_with_generator_like_cpp(
         &mut self,
         item_guid_generator: &wow_core::ObjectGuidGenerator,
@@ -23932,109 +22544,6 @@ impl WorldSession {
             criteria_tree_id,
         )
         .await;
-    }
-
-    pub(crate) fn enqueue_represented_quest_objective_progress_like_cpp(
-        &mut self,
-        event: RepresentedQuestObjectiveProgressEventLikeCpp,
-    ) {
-        self.represented_quest_objective_progress_events_like_cpp
-            .push_back(event);
-    }
-
-    pub(crate) async fn drain_represented_quest_objective_progress_with_generator_like_cpp(
-        &mut self,
-        item_guid_generator: &wow_core::ObjectGuidGenerator,
-    ) {
-        if self.represented_quest_objective_progress_draining_like_cpp {
-            return;
-        }
-
-        self.represented_quest_objective_progress_draining_like_cpp = true;
-        while let Some(event) = self
-            .represented_quest_objective_progress_events_like_cpp
-            .pop_front()
-        {
-            match event {
-                RepresentedQuestObjectiveProgressEventLikeCpp::MoneyChanged {
-                    old_money,
-                    new_money,
-                } => {
-                    self.update_represented_money_quest_objective_progress_like_cpp(
-                        item_guid_generator,
-                        old_money,
-                        new_money,
-                    )
-                    .await;
-                }
-                RepresentedQuestObjectiveProgressEventLikeCpp::CurrencyChanged {
-                    currency_id,
-                    change,
-                } => {
-                    let object_id = i32::try_from(currency_id).unwrap_or(i32::MAX);
-                    self.update_represented_currency_quest_objective_progress_like_cpp(
-                        item_guid_generator,
-                        currency_id,
-                        change,
-                    )
-                    .await;
-                    self.update_represented_storing_value_quest_objective_progress_like_cpp(
-                        item_guid_generator,
-                        QUEST_OBJECTIVE_HAVE_CURRENCY_LIKE_CPP,
-                        object_id,
-                        change,
-                        wow_core::ObjectGuid::new(0, 0),
-                    )
-                    .await;
-                    self.update_represented_storing_value_quest_objective_progress_like_cpp(
-                        item_guid_generator,
-                        QUEST_OBJECTIVE_OBTAIN_CURRENCY_LIKE_CPP,
-                        object_id,
-                        change,
-                        wow_core::ObjectGuid::new(0, 0),
-                    )
-                    .await;
-                }
-                RepresentedQuestObjectiveProgressEventLikeCpp::ReputationChanged {
-                    faction_id,
-                    change,
-                } => {
-                    let object_id = i32::try_from(faction_id).unwrap_or(i32::MAX);
-                    self.update_represented_reputation_quest_objective_progress_like_cpp(
-                        item_guid_generator,
-                        QUEST_OBJECTIVE_MIN_REPUTATION_LIKE_CPP,
-                        faction_id,
-                        change,
-                    )
-                    .await;
-                    self.update_represented_reputation_quest_objective_progress_like_cpp(
-                        item_guid_generator,
-                        QUEST_OBJECTIVE_MAX_REPUTATION_LIKE_CPP,
-                        faction_id,
-                        change,
-                    )
-                    .await;
-                    self.update_represented_storing_value_quest_objective_progress_like_cpp(
-                        item_guid_generator,
-                        QUEST_OBJECTIVE_INCREASE_REPUTATION_LIKE_CPP,
-                        object_id,
-                        change,
-                        wow_core::ObjectGuid::new(0, 0),
-                    )
-                    .await;
-                }
-            }
-        }
-        self.represented_quest_objective_progress_draining_like_cpp = false;
-    }
-
-    #[cfg(test)]
-    pub(crate) async fn drain_represented_quest_objective_progress_like_cpp(&mut self) {
-        let Some(generator) = self.item_guid_generator_like_cpp_for_bridge() else {
-            return;
-        };
-        self.drain_represented_quest_objective_progress_with_generator_like_cpp(generator.as_ref())
-            .await;
     }
 
     #[cfg(test)]
@@ -24323,64 +22832,6 @@ impl WorldSession {
             .await;
     }
 
-    /// Set the QuestXP store (loaded from QuestXP.db2).
-    pub fn set_quest_xp_store(&mut self, store: Arc<wow_data::quest_xp::QuestXpStore>) {
-        self.quest_xp_store = Some(store);
-    }
-
-    /// Set the QuestMoneyReward store (loaded from QuestMoneyReward.db2).
-    pub fn set_quest_money_reward_store(&mut self, store: Arc<QuestMoneyRewardStore>) {
-        self.quest_money_reward_store = Some(store);
-    }
-
-    fn set_loaded_quest_completed_bit_like_cpp(&mut self, quest_bit: u32) -> bool {
-        if quest_bit == 0 {
-            return false;
-        }
-
-        let field_offset = (quest_bit - 1) / QUESTS_COMPLETED_BITS_PER_BLOCK;
-        if field_offset as usize >= QUESTS_COMPLETED_BITS_SIZE {
-            return false;
-        }
-
-        let canonical_changed = self
-            .mutate_canonical_player_like_cpp(|player| {
-                player.set_quest_completed_bit_like_cpp(quest_bit, true)
-            })
-            .unwrap_or(false);
-        #[cfg(test)]
-        if self.player_handle_like_cpp.is_none() {
-            return self
-                .represented_quest_completed_bits_like_cpp
-                .insert(quest_bit);
-        }
-        canonical_changed
-    }
-
-    fn clear_loaded_quest_completed_bit_like_cpp(&mut self, quest_bit: u32) -> bool {
-        if quest_bit == 0 {
-            return false;
-        }
-
-        let field_offset = (quest_bit - 1) / QUESTS_COMPLETED_BITS_PER_BLOCK;
-        if field_offset as usize >= QUESTS_COMPLETED_BITS_SIZE {
-            return false;
-        }
-
-        let canonical_changed = self
-            .mutate_canonical_player_like_cpp(|player| {
-                player.set_quest_completed_bit_like_cpp(quest_bit, false)
-            })
-            .unwrap_or(false);
-        #[cfg(test)]
-        if self.player_handle_like_cpp.is_none() {
-            return self
-                .represented_quest_completed_bits_like_cpp
-                .remove(&quest_bit);
-        }
-        canonical_changed
-    }
-
     /// Set the player XP table (xp required per level).
     #[cfg(test)]
     pub fn set_player_xp_table(&mut self, table: Arc<Vec<u32>>) {
@@ -24447,10 +22898,6 @@ impl WorldSession {
 
     pub fn set_ffa_pvp_realm_like_cpp(&mut self, is_ffa_pvp_realm: bool) {
         self.is_ffa_pvp_realm_like_cpp = is_ffa_pvp_realm;
-    }
-
-    pub fn set_min_quest_scaled_xp_ratio_like_cpp(&mut self, ratio: u32) {
-        self.min_quest_scaled_xp_ratio_like_cpp = if ratio > 100 { 0 } else { ratio };
     }
 
     #[cfg(test)]
@@ -24541,76 +22988,6 @@ impl WorldSession {
     pub(crate) fn player_scaling_level_delta_like_cpp(&self) -> i32 {
         self.resolved_player_scaling_level_delta_like_cpp()
             .expect("test Player progression owner must resolve")
-    }
-
-    /// C++ `Player::GetQuestLevel`.
-    pub(crate) fn player_quest_level_like_cpp(
-        &self,
-        quest: &wow_data::quest::QuestTemplate,
-    ) -> i32 {
-        if quest.quest_level > 0 {
-            quest.quest_level
-        } else {
-            i32::from(self.player_level_like_cpp()).min(quest.quest_max_scaling_level)
-        }
-    }
-
-    /// Calculate XP reward for a quest.
-    /// C++ `Quest::XPValue(player, questLevel, xpDifficulty, xpMultiplier)`.
-    pub(crate) fn calculate_quest_xp(
-        &self,
-        difficulty: u32,
-        quest_level: i32,
-        xp_multiplier: f32,
-    ) -> u32 {
-        if let Some(store) = &self.quest_xp_store {
-            store.calculate_xp(
-                quest_level,
-                self.player_level_like_cpp(),
-                difficulty,
-                xp_multiplier,
-                self.min_quest_scaled_xp_ratio_like_cpp,
-            )
-        } else {
-            const XP_TABLE: [u32; 10] = [0, 50, 100, 200, 400, 650, 1000, 1500, 2500, 4000];
-            XP_TABLE[difficulty.min(9) as usize]
-        }
-    }
-
-    /// C++ `Player::GetQuestXPReward`.
-    pub(crate) fn quest_xp_reward_like_cpp(&self, quest: &wow_data::quest::QuestTemplate) -> u32 {
-        let Some(already_rewarded) = self.represented_player_has_rewarded_quest_like_cpp(quest.id)
-        else {
-            return 0;
-        };
-        if already_rewarded && !quest.is_df_quest_like_cpp() {
-            return 0;
-        }
-
-        self.calculate_quest_xp(
-            quest.reward_xp_difficulty,
-            self.player_quest_level_like_cpp(quest),
-            quest.reward_xp_multiplier,
-        )
-    }
-
-    /// C++ `Player::GetQuestMoneyReward` -> `Quest::MoneyValue`.
-    pub(crate) fn quest_money_reward_like_cpp(
-        &self,
-        quest: &wow_data::quest::QuestTemplate,
-    ) -> u32 {
-        let Some(store) = &self.quest_money_reward_store else {
-            return 0;
-        };
-        let quest_level = self.player_quest_level_like_cpp(quest).max(0) as u32;
-        let Some(row) = store.get(quest_level) else {
-            return 0;
-        };
-        let difficulty = quest.reward_money_difficulty as usize;
-        let Some(base) = row.difficulty.get(difficulty).copied() else {
-            return 0;
-        };
-        ((base as f32) * quest.reward_money_multiplier).round() as u32
     }
 
     /// Set the shared player registry (used for broadcast).
@@ -24934,262 +23311,11 @@ impl WorldSession {
                 .any(|registered| registered == prefix)
     }
 
-    pub(crate) fn load_seasonal_quest_status_like_cpp(
-        &mut self,
-        rows: impl IntoIterator<Item = SeasonalQuestStatusDbRowLikeCpp>,
-        quest_store: Option<&wow_data::quest::QuestStore>,
-        quest_v2_store: Option<&QuestV2Store>,
-    ) -> LoadSeasonalQuestStatusOutcomeLikeCpp {
-        // C++ Player::_LoadSeasonalQuestStatus clears first and always resets
-        // m_SeasonalQuestChanged at the end. Rust deliberately skips rows when
-        // no quest store is available, because C++ requires a real QuestTemplate.
-        let mut seasonal_quests = BTreeMap::<u16, BTreeMap<u32, u64>>::new();
-        let mut outcome = LoadSeasonalQuestStatusOutcomeLikeCpp::default();
-        for row in rows {
-            outcome.rows_seen += 1;
-
-            let Some(quest_store) = quest_store else {
-                outcome.skipped_no_quest_store += 1;
-                continue;
-            };
-
-            if quest_store.get(row.quest_id).is_none() {
-                outcome.skipped_missing_quest += 1;
-                continue;
-            }
-
-            let Ok(event_id) = u16::try_from(row.event_id) else {
-                outcome.skipped_event_out_of_range += 1;
-                continue;
-            };
-
-            let Ok(completed_time) = u64::try_from(row.completed_time) else {
-                // Bounded Rust safety divergence: C++ reads int64 then assigns
-                // to uint32, but Rust avoids turning negative DB data into a
-                // huge cooldown timestamp.
-                outcome.skipped_negative_completed_time += 1;
-                continue;
-            };
-
-            if seasonal_quests
-                .entry(event_id)
-                .or_default()
-                .insert(row.quest_id, completed_time)
-                .is_some()
-            {
-                outcome.replaced += 1;
-            } else {
-                outcome.inserted += 1;
-            }
-
-            let Some(quest_v2_store) = quest_v2_store else {
-                outcome.completed_bit_skipped_no_quest_v2_store += 1;
-                continue;
-            };
-
-            let quest_bit = quest_v2_store.get_quest_unique_bit_flag_like_cpp(row.quest_id);
-            if quest_bit == 0 {
-                outcome.completed_bit_skipped_zero_unique_bit += 1;
-                continue;
-            }
-
-            if self.set_loaded_quest_completed_bit_like_cpp(quest_bit) {
-                outcome.completed_bit_set += 1;
-            } else {
-                outcome.completed_bit_no_change_or_noop += 1;
-            }
-        }
-
-        let _ = self.mutate_player_quest_gameplay_like_cpp(|state| {
-            state.seasonal_quests = seasonal_quests;
-            state.seasonal_quest_changed = false;
-        });
-        outcome.seasonal_quest_changed = false;
-        outcome
-    }
-
-    pub(crate) fn reset_seasonal_quest_status_like_cpp(
-        &mut self,
-        event_id: u16,
-        event_start_time: u64,
-    ) -> ResetSeasonalQuestStatusOutcomeLikeCpp {
-        // C++ Player::ResetSeasonalQuestStatus: DB data deleted in caller.
-        if self
-            .mutate_player_quest_gameplay_like_cpp(|state| {
-                state.seasonal_quest_changed = false;
-            })
-            .is_none()
-        {
-            return ResetSeasonalQuestStatusOutcomeLikeCpp {
-                event_id,
-                event_start_time,
-                reason: ResetSeasonalQuestStatusReasonLikeCpp::MissingEvent,
-                removed_quest_ids: Vec::new(),
-                completed_bit_cleared: 0,
-                completed_bit_skipped_no_quest_v2_store: 0,
-                completed_bit_skipped_zero_unique_bit: 0,
-                completed_bit_no_change_or_noop: 0,
-                completed_bit_clear_unrepresented: 0,
-                event_bucket_erased: false,
-                seasonal_quest_changed: false,
-            };
-        }
-        let Some(mut recurrence) = self.player_quest_gameplay_snapshot_like_cpp() else {
-            return ResetSeasonalQuestStatusOutcomeLikeCpp {
-                event_id,
-                event_start_time,
-                reason: ResetSeasonalQuestStatusReasonLikeCpp::MissingEvent,
-                removed_quest_ids: Vec::new(),
-                completed_bit_cleared: 0,
-                completed_bit_skipped_no_quest_v2_store: 0,
-                completed_bit_skipped_zero_unique_bit: 0,
-                completed_bit_no_change_or_noop: 0,
-                completed_bit_clear_unrepresented: 0,
-                event_bucket_erased: false,
-                seasonal_quest_changed: false,
-            };
-        };
-        let Some(bucket) = recurrence.seasonal_quests.get_mut(&event_id) else {
-            return ResetSeasonalQuestStatusOutcomeLikeCpp {
-                event_id,
-                event_start_time,
-                reason: ResetSeasonalQuestStatusReasonLikeCpp::MissingEvent,
-                removed_quest_ids: Vec::new(),
-                completed_bit_cleared: 0,
-                completed_bit_skipped_no_quest_v2_store: 0,
-                completed_bit_skipped_zero_unique_bit: 0,
-                completed_bit_no_change_or_noop: 0,
-                completed_bit_clear_unrepresented: 0,
-                event_bucket_erased: false,
-                seasonal_quest_changed: false,
-            };
-        };
-
-        if bucket.is_empty() {
-            return ResetSeasonalQuestStatusOutcomeLikeCpp {
-                event_id,
-                event_start_time,
-                reason: ResetSeasonalQuestStatusReasonLikeCpp::EmptyEvent,
-                removed_quest_ids: Vec::new(),
-                completed_bit_cleared: 0,
-                completed_bit_skipped_no_quest_v2_store: 0,
-                completed_bit_skipped_zero_unique_bit: 0,
-                completed_bit_no_change_or_noop: 0,
-                completed_bit_clear_unrepresented: 0,
-                event_bucket_erased: false,
-                seasonal_quest_changed: false,
-            };
-        }
-
-        let removed_quest_ids: Vec<u32> = bucket
-            .iter()
-            .filter_map(|(quest_id, completed_time)| {
-                (*completed_time < event_start_time).then_some(*quest_id)
-            })
-            .collect();
-
-        for quest_id in &removed_quest_ids {
-            bucket.remove(quest_id);
-        }
-
-        let event_bucket_erased = bucket.is_empty() && !removed_quest_ids.is_empty();
-        if event_bucket_erased {
-            recurrence.seasonal_quests.remove(&event_id);
-        }
-        let seasonal_quests = recurrence.seasonal_quests;
-        let _ = self.mutate_player_quest_gameplay_like_cpp(|state| {
-            state.seasonal_quests = seasonal_quests;
-            state.seasonal_quest_changed = false;
-        });
-
-        let mut completed_bit_cleared = 0;
-        let mut completed_bit_skipped_no_quest_v2_store = 0;
-        let mut completed_bit_skipped_zero_unique_bit = 0;
-        let mut completed_bit_no_change_or_noop = 0;
-        for quest_id in &removed_quest_ids {
-            let Some(quest_v2_store) = self.quest_v2_store.as_ref().map(Arc::clone) else {
-                completed_bit_skipped_no_quest_v2_store += 1;
-                continue;
-            };
-
-            let quest_bit = quest_v2_store.get_quest_unique_bit_flag_like_cpp(*quest_id);
-            if quest_bit == 0 {
-                completed_bit_skipped_zero_unique_bit += 1;
-                continue;
-            }
-
-            if self.clear_loaded_quest_completed_bit_like_cpp(quest_bit) {
-                completed_bit_cleared += 1;
-            } else {
-                completed_bit_no_change_or_noop += 1;
-            }
-        }
-
-        ResetSeasonalQuestStatusOutcomeLikeCpp {
-            event_id,
-            event_start_time,
-            reason: if removed_quest_ids.is_empty() {
-                ResetSeasonalQuestStatusReasonLikeCpp::NoOlderCompletions
-            } else {
-                ResetSeasonalQuestStatusReasonLikeCpp::RemovedOlderCompletions
-            },
-            removed_quest_ids,
-            completed_bit_cleared,
-            completed_bit_skipped_no_quest_v2_store,
-            completed_bit_skipped_zero_unique_bit,
-            completed_bit_no_change_or_noop,
-            completed_bit_clear_unrepresented: 0,
-            event_bucket_erased,
-            seasonal_quest_changed: false,
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn seed_seasonal_quest_status_like_cpp(
-        &mut self,
-        event_id: u16,
-        quest_id: u32,
-        completed_time: u64,
-    ) {
-        let _ = self.mutate_player_quest_gameplay_like_cpp(|state| {
-            state
-                .seasonal_quests
-                .entry(event_id)
-                .or_default()
-                .insert(quest_id, completed_time);
-        });
-    }
-
     #[cfg(test)]
     pub(crate) fn seed_empty_seasonal_event_bucket_like_cpp(&mut self, event_id: u16) {
         let _ = self.mutate_player_quest_gameplay_like_cpp(|state| {
             state.seasonal_quests.entry(event_id).or_default();
         });
-    }
-
-    #[cfg(test)]
-    pub(crate) fn seasonal_quest_bucket_like_cpp(
-        &self,
-        event_id: u16,
-    ) -> Option<BTreeMap<u32, u64>> {
-        self.player_quest_gameplay_snapshot_like_cpp()?
-            .seasonal_quests
-            .get(&event_id)
-            .cloned()
-    }
-
-    #[cfg(test)]
-    pub(crate) fn set_seasonal_quest_changed_like_cpp_for_test(&mut self, changed: bool) {
-        let _ = self.mutate_player_quest_gameplay_like_cpp(|state| {
-            state.seasonal_quest_changed = changed;
-        });
-    }
-
-    #[cfg(test)]
-    pub(crate) fn seasonal_quest_changed_like_cpp(&self) -> bool {
-        self.player_quest_gameplay_snapshot_like_cpp()
-            .expect("test Player quest owner resolves")
-            .seasonal_quest_changed
     }
 
     /// Set the list of legitimate characters for this account.
@@ -27058,23 +25184,6 @@ impl WorldSession {
                 .cloned();
         }
         canonical.flatten()
-    }
-
-    #[cfg_attr(not(test), allow(unused_variables))]
-    pub(crate) fn record_represented_auction_replicate_request_like_cpp(
-        &mut self,
-        request: RepresentedAuctionReplicateRequestLikeCpp,
-    ) {
-        #[cfg(test)]
-        self.represented_auction_replicate_requests_like_cpp
-            .push(request);
-    }
-
-    #[cfg(test)]
-    pub(crate) fn represented_auction_replicate_requests_like_cpp(
-        &self,
-    ) -> &[RepresentedAuctionReplicateRequestLikeCpp] {
-        &self.represented_auction_replicate_requests_like_cpp
     }
 
     #[cfg_attr(not(test), allow(unused_variables))]
@@ -29461,15 +27570,6 @@ impl WorldSession {
             .is_some_and(|state| state.represented_status == Some(4))
     }
 
-    pub(crate) fn request_represented_battleground_leave_like_cpp(&mut self) {
-        #[cfg(test)]
-        {
-            self.represented_battleground_leave_requests_like_cpp = self
-                .represented_battleground_leave_requests_like_cpp
-                .saturating_add(1);
-        }
-    }
-
     fn unsummon_represented_pet_temporary_if_any_like_cpp(&mut self) {
         let Some(pet_guid) = self.player_pet_guid_state_like_cpp().flatten() else {
             return;
@@ -29897,11 +27997,6 @@ impl WorldSession {
     #[cfg(test)]
     pub(crate) fn resummon_pet_temporary_unsummoned_if_any_like_cpp(&mut self) {
         self.resummon_pet_temporary_unsummoned_like_cpp();
-    }
-
-    #[cfg(test)]
-    pub(crate) fn represented_battleground_leave_requests_like_cpp(&self) -> u32 {
-        self.represented_battleground_leave_requests_like_cpp
     }
 
     #[cfg_attr(not(test), allow(unused_variables))]
@@ -30507,63 +28602,6 @@ impl WorldSession {
         self.resolved_player_vitals_like_cpp().unwrap().1
     }
 
-    pub(crate) fn set_represented_resurrection_request_like_cpp(
-        &mut self,
-        request: PlayerResurrectionRequestLikeCpp,
-    ) -> bool {
-        let canonical = self
-            .with_owned_player_mut_like_cpp(|player| {
-                player.set_resurrection_request_like_cpp(request)
-            })
-            .is_some();
-        #[cfg(test)]
-        if canonical || self.player_handle_like_cpp.is_none() {
-            self.represented_resurrection_request_like_cpp = Some(request);
-        }
-        canonical || cfg!(test) && self.player_handle_like_cpp.is_none()
-    }
-
-    pub(crate) fn clear_represented_resurrection_request_like_cpp(&mut self) -> bool {
-        let canonical = self
-            .with_owned_player_mut_like_cpp(Player::clear_resurrection_request_like_cpp)
-            .is_some();
-        #[cfg(test)]
-        if canonical || self.player_handle_like_cpp.is_none() {
-            self.represented_resurrection_request_like_cpp = None;
-        }
-        canonical || cfg!(test) && self.player_handle_like_cpp.is_none()
-    }
-
-    pub(crate) fn represented_resurrection_requested_by_like_cpp(
-        &self,
-        resurrecter: ObjectGuid,
-    ) -> bool {
-        self.player_resurrection_state_snapshot_like_cpp()
-            .and_then(|state| state.request)
-            .is_some_and(|request| {
-                request.resurrecter != ObjectGuid::EMPTY && request.resurrecter == resurrecter
-            })
-    }
-
-    pub(crate) fn take_represented_resurrection_request_if_requested_by_like_cpp(
-        &mut self,
-        resurrecter: ObjectGuid,
-    ) -> Option<PlayerResurrectionRequestLikeCpp> {
-        let canonical = self
-            .with_owned_player_mut_like_cpp(|player| {
-                player.take_resurrection_request_if_requested_by_like_cpp(resurrecter)
-            })
-            .flatten();
-        #[cfg(test)]
-        if canonical.is_none() && self.player_handle_like_cpp.is_none() {
-            if !self.represented_resurrection_requested_by_like_cpp(resurrecter) {
-                return None;
-            }
-            return self.represented_resurrection_request_like_cpp.take();
-        }
-        canonical
-    }
-
     pub(crate) fn apply_represented_resurrection_health_like_cpp(&mut self, health: u32) {
         let Some((_, max_health, _)) = self.resolved_player_vitals_like_cpp() else {
             return;
@@ -30597,14 +28635,6 @@ impl WorldSession {
             });
         }
         canonical
-    }
-
-    #[cfg(test)]
-    pub(crate) fn represented_resurrection_request_like_cpp(
-        &self,
-    ) -> Option<PlayerResurrectionRequestLikeCpp> {
-        self.player_resurrection_state_snapshot_like_cpp()
-            .and_then(|state| state.request)
     }
 
     fn resolved_represented_total_stat_multiplier_for_stat_like_cpp(
@@ -31166,34 +29196,6 @@ impl WorldSession {
                     } if *effect_player_guid == player_guid && *stand_state == current_stand_state
                 )
             })
-    }
-
-    pub(crate) fn request_temporary_pet_unsummon_like_cpp(&mut self) {
-        self.invalidate_represented_character_pet_empty_authority_like_cpp();
-        #[cfg(test)]
-        {
-            self.temporary_pet_unsummon_requests_like_cpp = self
-                .temporary_pet_unsummon_requests_like_cpp
-                .saturating_add(1);
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn temporary_pet_unsummon_requests_like_cpp(&self) -> u32 {
-        self.temporary_pet_unsummon_requests_like_cpp
-    }
-
-    pub(crate) fn request_jump_proc_like_cpp(&mut self) {
-        #[cfg(test)]
-        {
-            self.movement_jump_proc_requests_like_cpp =
-                self.movement_jump_proc_requests_like_cpp.saturating_add(1);
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn movement_jump_proc_requests_like_cpp(&self) -> u32 {
-        self.movement_jump_proc_requests_like_cpp
     }
 
     fn represented_player_at_group_reward_distance_like_cpp(
@@ -32014,11 +30016,6 @@ impl WorldSession {
     }
 
     #[cfg(test)]
-    pub(crate) fn represented_duel_requests_like_cpp(&self) -> &[RepresentedDuelRequestedLikeCpp] {
-        &self.represented_duel_requests_like_cpp
-    }
-
-    #[cfg(test)]
     pub(crate) fn represented_force_deselects_like_cpp(
         &self,
     ) -> &[RepresentedForceDeselectLikeCpp] {
@@ -32097,23 +30094,6 @@ impl WorldSession {
             opponent_guid,
             SessionCommand::SendRepresentedDuelCountdownLikeCpp(
                 crate::session::mailbox::SendRepresentedDuelCountdownLikeCppCommand {
-                    packet_bytes,
-                },
-            ),
-        );
-    }
-
-    fn send_represented_duel_requested_to_opponent_like_cpp(
-        &self,
-        opponent_guid: ObjectGuid,
-        arbiter_guid: ObjectGuid,
-        packet_bytes: Vec<u8>,
-    ) {
-        self.try_send_connected_player_command_like_cpp(
-            opponent_guid,
-            SessionCommand::SendRepresentedDuelRequestedLikeCpp(
-                crate::session::mailbox::SendRepresentedDuelRequestedLikeCppCommand {
-                    arbiter_guid,
                     packet_bytes,
                 },
             ),
@@ -32388,72 +30368,10 @@ impl WorldSession {
         );
     }
 
-    pub(crate) fn represented_request_vehicle_exit_like_cpp(&mut self) -> bool {
-        let Some(seat_flags) = self
-            .player_vehicle_seat_state_like_cpp()
-            .and_then(|(flags, _)| flags)
-        else {
-            return false;
-        };
-        if !wow_data::vehicle_seat_flags_can_enter_or_exit_like_cpp(seat_flags) {
-            return false;
-        }
-
-        if !self.set_player_vehicle_seat_state_like_cpp(None, None) {
-            return false;
-        }
-        self.sync_player_registry_state_like_cpp();
-        true
-    }
-
     pub(crate) fn represented_current_vehicle_seat_can_switch_from_like_cpp(&self) -> bool {
         self.player_vehicle_seat_state_like_cpp()
             .and_then(|(flags, _)| flags)
             .is_some_and(wow_data::vehicle_seat_flags_can_switch_from_seat_like_cpp)
-    }
-
-    pub(crate) fn represented_request_adjacent_vehicle_seat_like_cpp(
-        &mut self,
-        next: bool,
-    ) -> bool {
-        match crate::handlers::vehicle::request_adjacent_vehicle_seat_action_like_cpp(
-            self.player_vehicle_seat_state_like_cpp()
-                .and_then(|(flags, _)| flags)
-                .is_some(),
-            self.represented_current_vehicle_seat_can_switch_from_like_cpp(),
-            next,
-        ) {
-            crate::handlers::vehicle::VehicleHandlerAction::ChangeSeat { seat_id, next } => {
-                #[cfg(test)]
-                self.represented_vehicle_seat_change_requests_like_cpp
-                    .push(RepresentedVehicleSeatChangeRequestLikeCpp { seat_id, next });
-                #[cfg(not(test))]
-                let _ = (seat_id, next);
-                true
-            }
-            _ => false,
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn represented_vehicle_seat_change_requests_like_cpp(
-        &self,
-    ) -> &[RepresentedVehicleSeatChangeRequestLikeCpp] {
-        &self.represented_vehicle_seat_change_requests_like_cpp
-    }
-
-    #[cfg(test)]
-    pub(crate) fn represented_vehicle_seat_spell_click_requests_like_cpp(
-        &self,
-    ) -> &[RepresentedVehicleSeatSpellClickRequestLikeCpp] {
-        &self.represented_vehicle_seat_spell_click_requests_like_cpp
-    }
-
-    #[cfg(test)]
-    pub(crate) fn represented_vehicle_enter_requests_like_cpp(
-        &self,
-    ) -> &[RepresentedVehicleEnterRequestLikeCpp] {
-        &self.represented_vehicle_enter_requests_like_cpp
     }
 
     fn represented_vehicle_base_guid_for_switch_like_cpp(&self) -> Option<ObjectGuid> {
@@ -32511,32 +30429,6 @@ impl WorldSession {
             }
             _ => false,
         }
-    }
-
-    pub(crate) fn represented_request_vehicle_switch_seat_like_cpp(
-        &mut self,
-        requested_vehicle: ObjectGuid,
-        seat_index: u8,
-    ) -> bool {
-        let Some(vehicle_base_guid) = self.represented_vehicle_base_guid_for_switch_like_cpp()
-        else {
-            return false;
-        };
-        let seat_id = seat_index as i8;
-        let requested_vehicle_exists_with_empty_seat = vehicle_base_guid != requested_vehicle
-            && self.represented_vehicle_seat_spell_click_plan_available_like_cpp(
-                requested_vehicle,
-                seat_id,
-            );
-        let action = crate::handlers::vehicle::request_vehicle_switch_seat_action_like_cpp(
-            true,
-            self.represented_current_vehicle_seat_can_switch_from_like_cpp(),
-            vehicle_base_guid,
-            requested_vehicle,
-            seat_index,
-            requested_vehicle_exists_with_empty_seat,
-        );
-        self.record_represented_vehicle_seat_action_like_cpp(action)
     }
 
     pub(crate) fn represented_ride_vehicle_interact_like_cpp(
@@ -33408,19 +31300,6 @@ impl WorldSession {
             }
             BattlePetAddOutcomeLikeCpp::Replayed(pet) => Ok(pet.guid),
         }
-    }
-
-    async fn battle_pet_add_request_committed_like_cpp(
-        &self,
-        request_key: BattlePetAddRequestKeyLikeCpp,
-    ) -> Result<bool, BattlePetAddFailureLikeCpp> {
-        let Some(attachment) = &self.battle_pet_account_attachment_like_cpp else {
-            return Ok(false);
-        };
-        attachment
-            .owner_like_cpp()
-            .add_request_committed_like_cpp(request_key)
-            .await
     }
 
     /// Publish one durable battle-pet addition exactly like the #160 session
@@ -34764,194 +32643,6 @@ impl WorldSession {
         true
     }
 
-    pub(crate) fn represented_quest_giver_involved_source_allows_quest_like_cpp(
-        &self,
-        source_guid: ObjectGuid,
-        quest_id: u32,
-        quest_store: &wow_data::quest::QuestStore,
-    ) -> bool {
-        if source_guid.is_any_type_creature() {
-            let Some(access) = self.represented_npc_can_interact_with_like_cpp(
-                source_guid,
-                NPCFlags1::QUEST_GIVER.bits(),
-                0,
-            ) else {
-                debug!(
-                    account = self.account_id,
-                    ?source_guid,
-                    quest_id,
-                    "QuestGiverCompleteReward: represented Creature source missing or not interactable"
-                );
-                return false;
-            };
-
-            return quest_store.creature_has_ender_relation_like_cpp(access.entry, quest_id);
-        }
-
-        if source_guid.is_game_object() {
-            let Some(access) =
-                self.represented_gameobject_questgiver_can_interact_with_like_cpp(source_guid)
-            else {
-                debug!(
-                    account = self.account_id,
-                    ?source_guid,
-                    quest_id,
-                    "QuestGiverCompleteReward: represented GameObject source missing or not interactable"
-                );
-                return false;
-            };
-
-            return quest_store.gameobject_has_ender_relation_like_cpp(access.entry, quest_id);
-        }
-
-        // Player and Item questgiver branches are not involved-quest sources in this represented slice.
-        // Match the C++ early-return shape by failing closed with no packet and no mutation.
-        debug!(
-            account = self.account_id,
-            ?source_guid,
-            quest_id,
-            "QuestGiverCompleteReward: unsupported represented source type"
-        );
-        false
-    }
-
-    pub(crate) fn represented_quest_giver_accept_source_allows_quest_like_cpp(
-        &self,
-        source_guid: ObjectGuid,
-        quest_id: u32,
-        quest_store: &wow_data::quest::QuestStore,
-    ) -> bool {
-        if source_guid.is_any_type_creature() {
-            let Some(access) = self.represented_npc_can_interact_with_like_cpp(
-                source_guid,
-                NPCFlags1::QUEST_GIVER.bits(),
-                0,
-            ) else {
-                debug!(
-                    account = self.account_id,
-                    ?source_guid,
-                    quest_id,
-                    "QuestGiverAcceptQuest: represented Creature source missing or not interactable"
-                );
-                return false;
-            };
-
-            return quest_store.creature_has_starter_relation_like_cpp(access.entry, quest_id);
-        }
-
-        if source_guid.is_game_object() {
-            let Some(access) =
-                self.represented_gameobject_questgiver_can_interact_with_like_cpp(source_guid)
-            else {
-                debug!(
-                    account = self.account_id,
-                    ?source_guid,
-                    quest_id,
-                    "QuestGiverAcceptQuest: represented GameObject source missing or not interactable"
-                );
-                return false;
-            };
-
-            return quest_store.gameobject_has_starter_relation_like_cpp(access.entry, quest_id);
-        }
-
-        // Player quest sharing and Item questgiver branches are not represented in this slice.
-        // Match the C++ early-return shape by failing closed with no packet and no mutation.
-        debug!(
-            account = self.account_id,
-            ?source_guid,
-            quest_id,
-            "QuestGiverAcceptQuest: unsupported represented source type"
-        );
-        false
-    }
-
-    pub(crate) fn send_represented_quest_giver_query_quest_like_cpp(
-        &mut self,
-        source_guid: ObjectGuid,
-        quest_id: u32,
-    ) -> bool {
-        let Some(quest_store) = self.quest_store.as_ref().map(Arc::clone) else {
-            debug!(
-                account = self.account_id,
-                ?source_guid,
-                quest_id,
-                "QuestGiverQueryQuest: missing quest store"
-            );
-            return false;
-        };
-        let Some(quest) = quest_store.get(quest_id).cloned() else {
-            warn!(
-                account = self.account_id,
-                ?source_guid,
-                quest_id,
-                "QuestGiverQueryQuest: unknown quest"
-            );
-            return false;
-        };
-
-        let menu_item = if source_guid.is_any_type_creature() {
-            let Some(access) = self.represented_npc_can_interact_with_like_cpp(
-                source_guid,
-                NPCFlags1::QUEST_GIVER.bits(),
-                0,
-            ) else {
-                debug!(
-                    account = self.account_id,
-                    ?source_guid,
-                    quest_id,
-                    "QuestGiverQueryQuest: represented Creature source missing or not interactable"
-                );
-                return false;
-            };
-
-            self.represented_quest_giver_query_creature_menu_item_like_cpp(
-                &quest_store,
-                access.entry,
-                &quest,
-            )
-        } else if source_guid.is_game_object() {
-            let Some(access) =
-                self.represented_gameobject_questgiver_can_interact_with_like_cpp(source_guid)
-            else {
-                debug!(
-                    account = self.account_id,
-                    ?source_guid,
-                    quest_id,
-                    "QuestGiverQueryQuest: represented GameObject source missing or not interactable"
-                );
-                return false;
-            };
-
-            self.represented_quest_giver_query_gameobject_menu_item_like_cpp(
-                &quest_store,
-                access.entry,
-                &quest,
-            )
-        } else {
-            debug!(
-                account = self.account_id,
-                ?source_guid,
-                quest_id,
-                "QuestGiverQueryQuest: unsupported represented source type"
-            );
-            return false;
-        };
-
-        let Some(menu_item) = menu_item else {
-            debug!(
-                account = self.account_id,
-                ?source_guid,
-                quest_id,
-                "QuestGiverQueryQuest: source has no represented starter/involved relation for quest"
-            );
-            return false;
-        };
-
-        self.send_represented_prepared_single_quest_like_cpp(source_guid, &menu_item, false);
-        true
-    }
-
     fn represented_quest_giver_query_creature_menu_item_like_cpp(
         &self,
         quest_store: &wow_data::quest::QuestStore,
@@ -35097,152 +32788,6 @@ impl WorldSession {
         })
     }
 
-    fn represented_player_quest_status_is_complete_or_incomplete_like_cpp(
-        &self,
-        quest_id: u32,
-    ) -> bool {
-        self.represented_player_quest_status_like_cpp(quest_id)
-            .is_some_and(|status| {
-                matches!(
-                    status,
-                    Some(
-                        crate::conditions::QUEST_STATUS_COMPLETE_LIKE_CPP
-                            | crate::conditions::QUEST_STATUS_INCOMPLETE_LIKE_CPP
-                    )
-                )
-            })
-    }
-
-    fn represented_player_quest_status_is_none_like_cpp(&self, quest_id: u32) -> bool {
-        self.represented_player_quest_status_like_cpp(quest_id) == Some(None)
-    }
-
-    fn send_represented_prepared_quest_like_cpp(
-        &mut self,
-        source_guid: ObjectGuid,
-        menu_items: Vec<RepresentedPreparedQuestMenuItemLikeCpp>,
-    ) {
-        if let [menu_item] = menu_items.as_slice() {
-            info!(
-                ?source_guid,
-                quest_id = menu_item.quest.id,
-                quest_icon = menu_item.quest_icon,
-                starter = menu_item.has_starter_relation,
-                involved = menu_item.has_involved_relation,
-                title = menu_item.quest.log_title.as_str(),
-                "Sending represented single prepared quest like C++"
-            );
-            self.send_represented_prepared_single_quest_like_cpp(source_guid, menu_item, true);
-            return;
-        }
-
-        info!(
-            ?source_guid,
-            quests = ?self.represented_quest_menu_item_log_rows_like_cpp(&menu_items),
-            "Sending represented QuestGiverQuestList like C++"
-        );
-        self.send_packet(&QuestGiverQuestList {
-            guid: source_guid,
-            greeting: String::new(),
-            greet_emote_delay: 0,
-            greet_emote_type: 0,
-            quests: menu_items
-                .iter()
-                .map(|item| self.quest_list_entry_from_menu_item_like_cpp(item))
-                .collect(),
-        });
-    }
-
-    fn send_represented_prepared_single_quest_like_cpp(
-        &mut self,
-        source_guid: ObjectGuid,
-        menu_item: &RepresentedPreparedQuestMenuItemLikeCpp,
-        auto_launched: bool,
-    ) {
-        let quest = &menu_item.quest;
-
-        if menu_item.quest_icon == QUEST_MENU_ICON_COMPLETE_LIKE_CPP {
-            let can_complete = self.can_reward_quest_represented_bounded_like_cpp(quest);
-            info!(
-                ?source_guid,
-                quest_id = quest.id,
-                can_complete,
-                auto_launched,
-                title = quest.log_title.as_str(),
-                "Sending represented QuestGiverRequestItems from complete icon like C++"
-            );
-            self.send_represented_quest_giver_request_items_with_completion_like_cpp(
-                source_guid,
-                quest,
-                can_complete,
-                auto_launched,
-            );
-            return;
-        }
-
-        if !menu_item.has_starter_relation && !menu_item.has_involved_relation {
-            // C++ would `SendCloseGossip()`. This represented seam has no close-gossip
-            // packet helper, and relation-derived menu items should not reach this branch.
-            return;
-        }
-
-        // C++ may auto-accept here; this represented-partial seam intentionally performs no
-        // quest log mutation and only chooses the outbound packet.
-        if quest.is_turn_in_like_cpp()
-            && quest.is_repeatable()
-            && !quest.is_daily_or_weekly_like_cpp()
-            && !quest.is_monthly_like_cpp()
-        {
-            info!(
-                ?source_guid,
-                quest_id = quest.id,
-                auto_launched,
-                title = quest.log_title.as_str(),
-                "Sending represented QuestGiverRequestItems for repeatable turn-in like C++"
-            );
-            let can_complete =
-                self.can_complete_repeatable_quest_represented_bounded_like_cpp(quest);
-            self.send_represented_quest_giver_request_items_with_completion_like_cpp(
-                source_guid,
-                quest,
-                can_complete,
-                auto_launched,
-            );
-        } else if quest.is_turn_in_like_cpp()
-            && !quest.is_daily_or_weekly_like_cpp()
-            && !quest.is_monthly_like_cpp()
-        {
-            let can_complete = self.can_reward_quest_represented_bounded_like_cpp(quest);
-            info!(
-                ?source_guid,
-                quest_id = quest.id,
-                can_complete,
-                auto_launched,
-                title = quest.log_title.as_str(),
-                "Sending represented QuestGiverRequestItems for turn-in like C++"
-            );
-            self.send_represented_quest_giver_request_items_with_completion_like_cpp(
-                source_guid,
-                quest,
-                can_complete,
-                auto_launched,
-            );
-        } else {
-            info!(
-                ?source_guid,
-                quest_id = quest.id,
-                auto_launched,
-                title = quest.log_title.as_str(),
-                "Sending represented QuestGiverQuestDetails like C++"
-            );
-            self.send_represented_quest_giver_quest_details_like_cpp(
-                source_guid,
-                quest,
-                auto_launched,
-            );
-        }
-    }
-
     fn represented_quest_menu_item_log_rows_like_cpp(
         &self,
         menu_items: &[RepresentedPreparedQuestMenuItemLikeCpp],
@@ -35261,182 +32806,6 @@ impl WorldSession {
                 )
             })
             .collect()
-    }
-
-    pub(crate) fn can_complete_repeatable_quest_represented_bounded_like_cpp(
-        &self,
-        quest: &wow_data::quest::QuestTemplate,
-    ) -> bool {
-        // C++ Player::CanCompleteRepeatableQuest is CanTakeQuest(false) && CanRewardQuest(false).
-        // The full CanRewardQuest blocker set includes disable checks, quest status, day/week/month/
-        // seasonal gates, level/skill/reputation, reward status, item/currency/money checks, etc.
-        // This bounded #611 seam must not overclaim completion while those blockers remain open.
-        if !self.can_take_quest(quest) {
-            return false;
-        }
-
-        let Some(inventory_item_counts) = self.represented_inventory_item_counts_like_cpp() else {
-            return false;
-        };
-        let mut saw_non_bound_item_objective = false;
-        for objective in &quest.objectives {
-            if objective.obj_type != QUEST_OBJECTIVE_ITEM_LIKE_CPP {
-                return false;
-            }
-
-            if (objective.flags2 & QUEST_OBJECTIVE_FLAG_2_QUEST_BOUND_ITEM_LIKE_CPP) != 0 {
-                continue;
-            }
-
-            saw_non_bound_item_objective = true;
-            let Ok(item_id) = u32::try_from(objective.object_id) else {
-                return false;
-            };
-            let Ok(required_count) = u32::try_from(objective.amount) else {
-                return false;
-            };
-            if inventory_item_counts.get(&item_id).copied().unwrap_or(0) < required_count {
-                return false;
-            }
-        }
-
-        // Item counts are represented only as partial evidence. Do not return true until the
-        // remaining C++ CanRewardQuest blockers are represented in this runtime path.
-        let _ = saw_non_bound_item_objective;
-        false
-    }
-
-    pub(crate) fn can_reward_quest_represented_bounded_like_cpp(
-        &self,
-        quest: &wow_data::quest::QuestTemplate,
-    ) -> bool {
-        // C++ `Player::CanRewardQuest(quest, false)` requires the quest to be complete
-        // unless it is a DF/turn-in-only quest, then rejects already rewarded
-        // non-repeatable quests. This represented seam covers the status/reward gates
-        // needed by `SendQuestGiverRequestItems`; inventory-capacity and long-tail
-        // reward blockers remain in the reward handler path.
-        if self.is_quest_disabled_like_cpp(quest.id) {
-            return false;
-        }
-        let Some(quests) = self.player_quest_gameplay_snapshot_like_cpp() else {
-            return false;
-        };
-
-        if !quest.is_df_quest_like_cpp()
-            && !quest.is_turn_in_like_cpp()
-            && !quests.statuses.get(&quest.id).is_some_and(|status| {
-                status.status == crate::conditions::QUEST_STATUS_COMPLETE_LIKE_CPP
-            })
-        {
-            return false;
-        }
-
-        if quests.rewarded_quest_ids.contains(&quest.id) && !quest.is_repeatable() {
-            return false;
-        }
-
-        true
-    }
-
-    pub(crate) fn send_represented_quest_giver_request_items_with_completion_like_cpp(
-        &mut self,
-        source_guid: ObjectGuid,
-        quest: &wow_data::quest::QuestTemplate,
-        can_complete: bool,
-        auto_launched: bool,
-    ) {
-        let collect = quest
-            .objectives
-            .iter()
-            .filter(|objective| objective.obj_type == QUEST_OBJECTIVE_ITEM_LIKE_CPP)
-            .map(|objective| QuestGiverRequestItemsCollect {
-                object_id: objective.object_id,
-                amount: objective.amount,
-                flags: objective.flags,
-            })
-            .collect::<Vec<_>>();
-        let currency = quest
-            .objectives
-            .iter()
-            .filter(|objective| objective.obj_type == QUEST_OBJECTIVE_CURRENCY_LIKE_CPP)
-            .map(|objective| QuestGiverRequestItemsCurrency {
-                currency_id: objective.object_id,
-                amount: objective.amount,
-            })
-            .collect::<Vec<_>>();
-        let money_to_get = quest
-            .objectives
-            .iter()
-            .filter(|objective| objective.obj_type == QUEST_OBJECTIVE_MONEY_LIKE_CPP)
-            .map(|objective| objective.amount)
-            .sum::<i32>();
-        self.send_packet(&QuestGiverRequestItems {
-            giver_guid: source_guid,
-            giver_creature_id: quest_giver_creature_id_from_source_like_cpp(source_guid),
-            quest_id: quest.id,
-            comp_emote_delay: 0,
-            comp_emote_type: 0,
-            quest_flags: [quest.flags, quest.flags_ex, quest.flags_ex2],
-            suggested_party_members: quest.suggested_group_num,
-            money_to_get,
-            collect,
-            currency,
-            status_flags: if can_complete { 0xFF } else { 0xFD },
-            title: quest.log_title.clone(),
-            completion_text: quest.area_description.clone(),
-            auto_launched,
-        });
-    }
-
-    pub(crate) fn send_represented_quest_giver_offer_reward_like_cpp(
-        &mut self,
-        source_guid: ObjectGuid,
-        quest: &wow_data::quest::QuestTemplate,
-        auto_launched: bool,
-    ) {
-        self.send_packet(&QuestGiverOfferReward {
-            giver_guid: source_guid,
-            giver_creature_id: quest_giver_creature_id_from_source_like_cpp(source_guid),
-            quest_id: quest.id,
-            quest_flags: [quest.flags, quest.flags_ex, quest.flags_ex2],
-            suggested_party_members: quest.suggested_group_num,
-            rewards: quest_rewards_block_like_cpp(quest),
-            title: quest.log_title.clone(),
-            reward_text: quest.quest_completion_log.clone(),
-            auto_launched,
-        });
-    }
-
-    pub(crate) fn send_represented_quest_giver_quest_details_like_cpp(
-        &mut self,
-        source_guid: ObjectGuid,
-        quest: &wow_data::quest::QuestTemplate,
-        auto_launched: bool,
-    ) {
-        let objectives: Vec<QuestObjectiveSimple> = quest
-            .objectives
-            .iter()
-            .map(|obj| QuestObjectiveSimple {
-                id: obj.id,
-                object_id: obj.object_id,
-                amount: obj.amount,
-                obj_type: obj.obj_type,
-            })
-            .collect();
-
-        self.send_packet(&QuestGiverQuestDetails {
-            giver_guid: source_guid,
-            giver_creature_id: quest_giver_creature_id_from_source_like_cpp(source_guid),
-            quest_id: quest.id,
-            quest_flags: [quest.flags, quest.flags_ex, quest.flags_ex2],
-            suggested_party_members: quest.suggested_group_num,
-            objectives,
-            rewards: quest_rewards_block_like_cpp(quest),
-            title: quest.log_title.clone(),
-            description: quest.quest_description.clone(),
-            log_description: quest.log_description.clone(),
-            auto_launched,
-        });
     }
 
     pub(crate) fn represented_creature_gossip_text_like_cpp(
@@ -36051,21 +33420,6 @@ impl WorldSession {
         &self.cuf_profiles_like_cpp
     }
 
-    #[cfg(test)]
-    pub(crate) fn movement_visibility_refresh_requests_like_cpp(&self) -> u32 {
-        self.movement_visibility_refresh_requests_like_cpp
-    }
-
-    pub(crate) fn consume_movement_visibility_refresh_request_like_cpp(&mut self) -> bool {
-        if self.movement_visibility_refresh_requests_like_cpp == 0 {
-            return false;
-        }
-        self.movement_visibility_refresh_requests_like_cpp = self
-            .movement_visibility_refresh_requests_like_cpp
-            .saturating_sub(1);
-        true
-    }
-
     pub(crate) fn player_taxi_state_snapshot_like_cpp(
         &self,
     ) -> Option<wow_entities::PlayerTaxiState> {
@@ -36145,13 +33499,6 @@ impl WorldSession {
             .push(request);
     }
 
-    #[cfg(test)]
-    pub(crate) fn represented_activate_taxi_requests_like_cpp(
-        &self,
-    ) -> &[RepresentedActivateTaxiLikeCpp] {
-        &self.represented_activate_taxi_requests_like_cpp
-    }
-
     #[cfg_attr(not(test), allow(unused_variables))]
     pub(crate) fn record_represented_confirm_barbers_choice_like_cpp(
         &mut self,
@@ -36162,13 +33509,6 @@ impl WorldSession {
             self.represented_confirm_barbers_choice_requests_like_cpp
                 .push(request);
         }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn represented_confirm_barbers_choice_requests_like_cpp(
-        &self,
-    ) -> &[RepresentedConfirmBarbersChoiceLikeCpp] {
-        &self.represented_confirm_barbers_choice_requests_like_cpp
     }
 
     #[cfg_attr(not(test), allow(unused_variables))]
@@ -36258,34 +33598,10 @@ impl WorldSession {
     }
 
     #[cfg(test)]
-    pub(crate) fn represented_confirm_respec_wipe_requests_like_cpp(
-        &self,
-    ) -> &[RepresentedConfirmRespecWipeLikeCpp] {
-        &self.represented_confirm_respec_wipe_requests_like_cpp
-    }
-
-    #[cfg(test)]
     pub(crate) fn represented_talent_respec_criteria_events_like_cpp(
         &self,
     ) -> &[RepresentedTalentRespecCriteriaEventLikeCpp] {
         &self.represented_talent_respec_criteria_events_like_cpp
-    }
-
-    #[cfg_attr(not(test), allow(unused_variables))]
-    pub(crate) fn record_represented_adventure_map_start_quest_like_cpp(
-        &mut self,
-        request: RepresentedAdventureMapStartQuestLikeCpp,
-    ) {
-        #[cfg(test)]
-        self.represented_adventure_map_start_quest_requests_like_cpp
-            .push(request);
-    }
-
-    #[cfg(test)]
-    pub(crate) fn represented_adventure_map_start_quest_requests_like_cpp(
-        &self,
-    ) -> &[RepresentedAdventureMapStartQuestLikeCpp] {
-        &self.represented_adventure_map_start_quest_requests_like_cpp
     }
 
     pub(crate) fn resolved_is_in_taxi_flight_like_cpp(&self) -> Option<bool> {
@@ -36406,11 +33722,6 @@ impl WorldSession {
             self.player_pvp_enabled_like_cpp = pvp_enabled;
             self.player_in_pvp_flag_like_cpp = in_pvp_flag;
         }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn temporary_pet_resummon_requests_like_cpp(&self) -> u32 {
-        self.temporary_pet_resummon_requests_like_cpp
     }
 
     #[cfg(test)]
@@ -36944,111 +34255,6 @@ impl WorldSession {
         false
     }
 
-    pub(crate) fn set_represented_pending_quest_sharing_like_cpp(
-        &mut self,
-        sender_guid: ObjectGuid,
-        quest_id: u32,
-    ) {
-        #[cfg(test)]
-        if self.player_handle_like_cpp.is_none() {
-            self.represented_pending_quest_sharing_like_cpp =
-                Some(RepresentedPendingQuestSharingLikeCpp {
-                    sender_guid,
-                    quest_id,
-                });
-            self.sync_player_registry_state_like_cpp();
-            return;
-        }
-        let _ = self.mutate_player_quest_gameplay_like_cpp(|state| {
-            state.pending_share = Some((sender_guid, quest_id));
-        });
-        self.sync_player_registry_state_like_cpp();
-    }
-
-    pub(crate) fn clear_represented_pending_quest_sharing_like_cpp(&mut self) {
-        #[cfg(test)]
-        if self.player_handle_like_cpp.is_none() {
-            self.represented_pending_quest_sharing_like_cpp = None;
-            self.sync_player_registry_state_like_cpp();
-            return;
-        }
-        let _ = self.mutate_player_quest_gameplay_like_cpp(|state| {
-            state.pending_share = None;
-        });
-        self.sync_player_registry_state_like_cpp();
-    }
-
-    pub(crate) fn represented_pending_quest_sharing_like_cpp(
-        &self,
-    ) -> Option<RepresentedPendingQuestSharingLikeCpp> {
-        #[cfg(test)]
-        if self.player_handle_like_cpp.is_none() {
-            return self.represented_pending_quest_sharing_like_cpp;
-        }
-        self.player_quest_gameplay_snapshot_like_cpp()
-            .and_then(|state| state.pending_share)
-            .map(
-                |(sender_guid, quest_id)| RepresentedPendingQuestSharingLikeCpp {
-                    sender_guid,
-                    quest_id,
-                },
-            )
-    }
-
-    pub(crate) fn set_represented_daily_quest_completed_like_cpp_for_test(
-        &mut self,
-        quest_id: u32,
-        completed: bool,
-    ) {
-        let _ = self.mutate_player_quest_gameplay_like_cpp(|state| {
-            if completed {
-                state.daily_quest_ids.insert(quest_id);
-            } else {
-                state.daily_quest_ids.remove(&quest_id);
-            }
-        });
-        self.sync_player_registry_state_like_cpp();
-    }
-
-    pub(crate) fn set_represented_df_quest_like_cpp_for_test(
-        &mut self,
-        quest_id: u32,
-        present: bool,
-    ) {
-        let _ = self.mutate_player_quest_gameplay_like_cpp(|state| {
-            if present {
-                state.df_quest_ids.insert(quest_id);
-            } else {
-                state.df_quest_ids.remove(&quest_id);
-            }
-        });
-        self.sync_player_registry_state_like_cpp();
-    }
-
-    #[cfg(test)]
-    pub(crate) fn represented_timed_quest_removals_like_cpp(&self) -> &[u32] {
-        &self.represented_timed_quest_removals_like_cpp
-    }
-
-    #[cfg(test)]
-    pub(crate) fn represented_quest_reward_skill_updates_like_cpp(&self) -> &[(u32, u32)] {
-        &self.represented_quest_reward_skill_updates_like_cpp
-    }
-
-    #[cfg(test)]
-    pub(crate) fn represented_quest_reward_spell_casts_like_cpp(
-        &self,
-    ) -> &[RepresentedQuestRewardSpellCastLikeCpp] {
-        &self.represented_quest_reward_spell_casts_like_cpp
-    }
-
-    #[cfg(test)]
-    pub(crate) fn represented_quest_reward_titles_like_cpp(
-        &self,
-    ) -> &[RepresentedQuestRewardTitleLikeCpp] {
-        &self.represented_quest_reward_titles_like_cpp
-    }
-
     pub(crate) fn represented_learn_title_like_cpp(&mut self, title_id: u32) {
         let _canonical = self
             .with_owned_player_mut_like_cpp(|player| player.learn_title_like_cpp(title_id))
@@ -37122,109 +34328,6 @@ impl WorldSession {
             return self.represented_chosen_title_like_cpp;
         }
         canonical.expect("test Player title owner must resolve")
-    }
-
-    #[cfg(test)]
-    pub(crate) fn represented_quest_reward_talent_points_like_cpp(
-        &self,
-    ) -> &[RepresentedQuestRewardTalentPointsLikeCpp] {
-        &self.represented_quest_reward_talent_points_like_cpp
-    }
-
-    #[cfg(test)]
-    pub(crate) fn represented_quest_reward_mails_like_cpp(
-        &self,
-    ) -> &[RepresentedQuestRewardMailLikeCpp] {
-        &self.represented_quest_reward_mails_like_cpp
-    }
-
-    #[cfg(test)]
-    pub(crate) fn represented_quest_reward_reputations_like_cpp(
-        &self,
-    ) -> &[RepresentedQuestRewardReputationLikeCpp] {
-        &self.represented_quest_reward_reputations_like_cpp
-    }
-
-    #[cfg(test)]
-    pub(crate) fn represented_quest_push_result_responses_like_cpp(
-        &self,
-    ) -> &[RepresentedQuestPushResultResponseLikeCpp] {
-        &self.represented_quest_push_result_responses_like_cpp
-    }
-
-    #[cfg(test)]
-    pub(crate) fn represented_quest_push_result_sender_mismatch_count_like_cpp(&self) -> u32 {
-        self.represented_quest_push_result_sender_mismatch_count_like_cpp
-    }
-
-    pub(crate) fn record_represented_quest_push_result_response_like_cpp(
-        &mut self,
-        response: RepresentedQuestPushResultResponseLikeCpp,
-    ) {
-        #[cfg(test)]
-        self.represented_quest_push_result_responses_like_cpp
-            .push(response);
-        #[cfg(not(test))]
-        let _ = response;
-    }
-
-    pub(crate) fn record_represented_quest_push_result_sender_mismatch_like_cpp(&mut self) {
-        #[cfg(test)]
-        {
-            self.represented_quest_push_result_sender_mismatch_count_like_cpp = self
-                .represented_quest_push_result_sender_mismatch_count_like_cpp
-                .saturating_add(1);
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn represented_quest_confirm_accepts_like_cpp(
-        &self,
-    ) -> &[RepresentedQuestConfirmAcceptLikeCpp] {
-        &self.represented_quest_confirm_accepts_like_cpp
-    }
-
-    pub(crate) fn record_represented_quest_confirm_accept_like_cpp(
-        &mut self,
-        evidence: RepresentedQuestConfirmAcceptLikeCpp,
-    ) {
-        #[cfg(test)]
-        self.represented_quest_confirm_accepts_like_cpp
-            .push(evidence);
-        #[cfg(not(test))]
-        let _ = evidence;
-    }
-
-    pub(crate) fn represented_quest_complete_status_updates_like_cpp(
-        &self,
-    ) -> &[RepresentedQuestCompleteStatusUpdateLikeCpp] {
-        &self.represented_quest_complete_status_updates_like_cpp
-    }
-
-    pub(crate) fn record_represented_quest_complete_status_update_like_cpp(
-        &mut self,
-        evidence: RepresentedQuestCompleteStatusUpdateLikeCpp,
-    ) {
-        self.represented_quest_complete_status_updates_like_cpp
-            .push(evidence);
-    }
-
-    #[cfg(test)]
-    pub(crate) fn represented_push_quest_to_party_outcomes_like_cpp(
-        &self,
-    ) -> &[RepresentedPushQuestToPartyOutcomeLikeCpp] {
-        &self.represented_push_quest_to_party_outcomes_like_cpp
-    }
-
-    pub(crate) fn record_represented_push_quest_to_party_outcome_like_cpp(
-        &mut self,
-        outcome: RepresentedPushQuestToPartyOutcomeLikeCpp,
-    ) {
-        #[cfg(test)]
-        self.represented_push_quest_to_party_outcomes_like_cpp
-            .push(outcome);
-        #[cfg(not(test))]
-        let _ = outcome;
     }
 
     /// Get the logged-in player GUID.
